% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_absolute.R
\name{absolute_disease_map}
\alias{absolute_disease_map}
\title{ABSOLUTE Disease Map}
\usage{
absolute_disease_map()
}
\description{
A helper function, which just return the disease map for
ABSOLUTE algorithm.
}
