% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anno_hub.R
\name{anno_hub}
\alias{anno_hub}
\title{Get AnnotationHub object}
\usage{
anno_hub(
  ...,
  version = NULL,
  hub = NULL,
  cache = NULL,
  proxy = NULL,
  localHub = NULL
)
}
\arguments{
\item{...}{See \link[AnnotationHub:AnnotationHub-class]{AnnotationHub} for additional
arguments.}

\item{version}{A valid Bioconductor version, see:
\url{https://bioconductor.org/config.yaml} for all version.
\link[AnnotationHub:AnnotationHub-class]{AnnotationHub} restrict the AnnotationHub
snapshotDate before the Bioconductor version, this function just bypass the
restriction to let you query all AnnotationHub data.}

\item{hub}{The URL for the online AnnotationHub.}

\item{cache}{The file system location of the local AnnotationHub cache.}

\item{proxy}{Set the proxy.}

\item{localHub}{A bool, whether to use the cache.}
}
\value{
A \link[AnnotationHub:AnnotationHub-class]{AnnotationHub} instance.
}
\description{
Get AnnotationHub object
}
