% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_Agilent_url.R
\name{annotation_Agilent_url}
\alias{annotation_Agilent_url}
\title{Agilent annotation url}
\usage{
annotation_Agilent_url()
}
\value{
URL
}
\description{
Agilent annotation url
}
\examples{
annotation_Agilent_url()
}
