% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio2df.R
\name{bio2df}
\alias{bio2df}
\alias{bio2df.SummarizedExperiment}
\title{Transfrom a bioconductor Object into a data.frame}
\usage{
bio2df(x, ...)

\method{bio2df}{SummarizedExperiment}(x, ..., assay = NULL, row_as_observations = FALSE)
}
\arguments{
\item{x}{A \link[SummarizedExperiment:RangedSummarizedExperiment-class]{SummarizedExperiment}
or \link[Biobase:class.ExpressionSet]{ExpressionSet} object.}

\item{...}{For \code{bio2df}, other arguments passed to specific methods.}

\item{assay}{Integer scalar or string indicating which assay of \code{x} to use.}

\item{row_as_observations}{A scalar logical value indicates whether use row
as the observations, if TRUE, a data.frame combined the assay data and
\code{colData(x)} will be returned, otherwise, a data.frame combined the
transposed assay data and \code{rowData(x)}.}
}
\value{
A data.frame.
}
\description{
It's often we want to combine the expression data and phenotypic (or feature)
data into a data.frame to do something like plot (ggplot2) or simple
statistical test.
}
