% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomisc-package.R
\docType{package}
\name{biomisc-package}
\alias{biomisc}
\alias{biomisc-package}
\title{biomisc: Miscellaneous functions for Bioinformatics}
\description{
Miscellaneous functions for Bioinformatics.
}
\author{
\strong{Maintainer}: Yun Peng \email{yunyunpp96@outlook.com}

}
\keyword{internal}
