% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellmarker.R
\name{cellmarker_search}
\alias{cellmarker_search}
\title{Seach CellMarker database
Details see http://xteam.xbio.top/CellMarker/index.jsp}
\usage{
cellmarker_search(markers, species = "human", internal = NULL)
}
\arguments{
\item{markers}{the markers to search in the CellMarker database, should be
the gene symbol.}

\item{species}{a scalar string, "human" or "mouse".}

\item{internal}{logical value, indicates whether to use internal CellMarker
data or not. \code{NULL} indicates this will be determined by the CellMarker has
been downloaded once, if we have used this function once with a \code{internal}
FALSE, the value \code{NULL} will indicate \code{TRUE}. The internal data is downloaded
from CellMarker (2022-12-04).}
}
\description{
Seach CellMarker database
Details see http://xteam.xbio.top/CellMarker/index.jsp
}
