% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnmf.R
\name{cnmf}
\alias{cnmf}
\title{Modified cNMF}
\usage{
cnmf(
  matrix,
  min_fraction = 0.002,
  k = 15L,
  n_iters = 100L,
  local_neighborhood_size = 0.3,
  min_dist = 0.03,
  numb_smp = NULL
)
}
\arguments{
\item{matrix}{A matrix with row is features and column is samples.}

\item{min_fraction}{Only featuers with value (above 0) in at least
\code{min_fraction} of samples are used in cNMF.}

\item{k}{Number of programs.}

\item{n_iters}{The number of times to run NMF internally before making the
consensus}

\item{local_neighborhood_size}{Determines number of neighbors to use for
calculating KNN distance as local_neighborhood_size * n_iters.}

\item{min_dist}{The distance used for filtering out outliers}

\item{numb_smp}{Number of cells to select, can be a floating number between 0
and 1, which means the fraction of cells to choose (if NULL or <=0 no
downsampling will be applied).}
}
\value{
A list.
}
\description{
A modified version of the cNMF algorithm, implemented in R
}
\references{
Simmons, S.K., Lithwick-Yanai, G., Adiconis, X. et al. Mostly natural
sequencing-by-synthesis for scRNA-seq using Ultima sequencing. Nat Biotechnol
41, 204–211 (2023). https://doi.org/10.1038/s41587-022-01452-6
}
\seealso{
\itemize{
\item \url{https://github.com/seanken/CompareSequence}
\item \url{https://github.com/dylkot/cNMF/blob/master/src/cnmf/cnmf.py}
}
}
