% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus_modules.R
\name{consensus_modules}
\alias{consensus_modules}
\title{Identify consensus modules.}
\usage{
consensus_modules(
  module_list,
  cluster_fn = igraph::cluster_infomap,
  ...,
  v_min = 2L,
  s_min = 3L,
  threshold = 5L
)
}
\arguments{
\item{module_list}{A named list (defining tumor type) of named sub-list of
atomic characters (each stands for the identified modules in specific
tumor). The names of the list must be unique and blank names are not
allowed.}

\item{cluster_fn}{Function used to find communities (ususally function starts
with "cluster_" in igraph package). Always accept a "graph" object.}

\item{...}{Other argument passed to \code{cluster_fn}.}

\item{v_min}{Gene–gene connections were filtered out if they occurred in
fewer than \code{v_min} individual tumor modules.}

\item{s_min}{genes connected with fewer than \code{s_min} genes were removed.}

\item{threshold}{Indicates the minimum number of genes required to filter a
module, failing which, it shall not be considered.}
}
\value{
A Consensus_module list of filtered communities by \code{threshold}, each
identified by their members, with \code{raw_modules}, \code{full_adj} and \code{adj_list}
attached in the attributes.
}
\description{
Identify consensus modules.
}
\references{
Barkley, D., Moncada, R., Pour, M. et al. Cancer cell states recur across
tumor types and form specific interactions with the tumor microenvironment.
Nat Genet 54, 1192–1201 (2022). https://doi.org/10.1038/s41588-022-01141-9
}
\seealso{
\url{https://github.com/yanailab/PanCancer}
}
