% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context_sizefactor.R
\name{context_sizefactor}
\alias{context_sizefactor}
\title{Evaluate the sizefactor to transform context between different scale}
\usage{
context_sizefactor(
  context,
  method = c("genome", "exome", "exome2genome", "genome2exome"),
  ref_genome = NULL,
  exome_ranges = NULL,
  contigs = NULL
)
}
\arguments{
\item{context}{A character of nucleotide sequence, all must have the same
number of characters.}

\item{method}{One of "genome", "exome", "exome2genome", "genome2exome", If
the input mut_data only contains the counts of the mutations observed in
each context, then the mut_data must be normalized.
\itemize{
\item If method is set to 'exome', the input data frame is normalized by number
of times each trinucleotide context is observed in the exome.
\item If method is set to 'genome', the input data frame is normalized by number
of times each trinucleotide context is observed in the genome.
\item If method is set to 'exome2genome', normalization is performed to reflect
the absolute frequency of each trinucleotide context as it would occur
across the whole genome. Thus the count data for each trinucleotide context
is multiplied by a ratio of that trinucleotide's occurence in the genome to
the trinucleotide's occurence in the exome.

Details see \url{https://github.com/raerose01/deconstructSigs}
}}

\item{ref_genome}{BSgenome object or name of the installed BSgenome package.
Default: "hg19". Details see "genome" argument of
\link[BSgenome:available.genomes]{getBSgenome}.}

\item{exome_ranges}{A \link[GenomicRanges:GRanges-class]{GenomicRanges} Object define
exome ranges.}

\item{contigs}{An atomic vector specifying the chromosome to define the  size
factor to normalize the context matrix.}
}
\value{
A numeric size factor for each mut_context
}
\description{
Evaluate the sizefactor to transform context between different scale
}
