% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_geo_series_matrix.R
\name{download_geo_series_matrix}
\alias{download_geo_series_matrix}
\title{Download geo series matrix}
\usage{
download_geo_series_matrix(
  ...,
  chrome = FALSE,
  sleep = 5,
  browser = getOption("browser")
)
}
\arguments{
\item{...}{one or more gse}

\item{chrome}{logical, whether to download by chrome}

\item{sleep}{wait second, default is 5}

\item{browser}{browser}
}
\value{
series matrix url
}
\description{
Download geo series matrix
}
