% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigengenes.R
\name{eigengenes}
\alias{eigengenes}
\title{Create a module eigengene matrix}
\usage{
eigengenes(dat, geneSets)
}
\arguments{
\item{dat}{An expression matrix or matrix-like object, with rows corresponding to
probes and columns to samples. Data is presumed to be filtered and normalized
prior to dimensionality reduction. For count data, this means undergoing some
sort of variance stabilizing transformation, such as \code{\link{lcpm}, 
\link[DESeq2]{vst}, or \link[DESeq2]{rlog}}.}

\item{geneSets}{A named list of one or several gene sets.}
}
\description{
This function reduces an expression matrix to the module level.
}
\details{
Eigengenes are a convenient way to summarize data from a given gene set into a 
single vector of length equal to the study sample size. The eigengene of a module 
represents the first principle component of the sample by probe expression matrix 
for all genes of which the module is composed. This is useful for correlating gene 
sets with one another, or associating them with clinical variables of interest. 
They may even be used to create a meta-network of module eigengenes.
}
\examples{
# Simulate data
mat <- matrix(rnorm(5000 * 10), nrow = 5000, ncol = 10)
grp <- rep(c("A", "B"), each = 5)
geneSets = list()
for (i in 0:10) {
  genes <- ((30 * i) + 1):(30 * (i + 1))
  mat[genes, grp == "A"] <- mat[genes, grp == "A"] + rnorm(1)
  geneSets[[paste("Set", i)]] <- genes
}

# Create eigengene matrix
eg_mat <- eigengenes(mat, geneSets)

}
\references{
Zhang, B. & Horvath, S. (2005). "A General Framework for Weighted Gene 
Co-Expression Network Analysis". \emph{Stat. Appl. Genet. Molec. Biol.}, \emph{4}: 
1, 17.
\url{https://www.ncbi.nlm.nih.gov/pubmed/16646834}

Langfelder, P. & Horvath, S. (2007). "Eigengene networks for studying the 
relationships between co-expression modules." \emph{BMC Bioinformatics}, \emph{1}:
54.
\url{http://bmcsystbiol.biomedcentral.com/articles/10.1186/1752-0509-1-54}

Horvath S. & Dong, J. (2008). "Geometric Interpretation of Gene Coexpression Network 
Analysis." \emph{PLoS Comput. Biol.}, \emph{4}(8): e1000117.
\url{http://journals.plos.org/ploscompbiol/article?id=10.1371/journal.pcbi.1000117}
}
