% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_ccf.R
\name{estimate_ccf}
\alias{estimate_ccf}
\title{Estimating CCF (absolute and phylogenetic)}
\usage{
estimate_ccf(
  mut_cn_data,
  subclone_metric = "subclone_prop",
  min_subclonal = NULL,
  subclone_correction = FALSE,
  pvalue_correction = NULL,
  min_vaf_to_explain = NULL
)
}
\arguments{
\item{mut_cn_data}{A data.frame with mutation and copy number data.}

\item{subclone_metric}{A string, "ccf" or "subclone_prop" specifying how to
choose subclone.}

\item{min_subclonal}{Minimal copy number to define subclone.}

\item{subclone_correction}{A scalar logical indicates whether subclonal copy
number correction be used.}

\item{pvalue_correction}{The method of multiple testing correction be applied
for the copy number correcting mutations. If NULL, no multiple testing
correction will be used.}

\item{min_vaf_to_explain}{A numeric, the minimal vaf value to define
subclone.}
}
\description{
For CONIPHER anlayis, use subclone_metric = "ccf", min_subclonal = 0.05,
subclone_correction = TRUE, pvalue_correction = "BH", min_vaf_to_explain =
0.05.
}
