% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_ftp.suppl.URL.R
\name{geo_ftp.suppl.URL}
\alias{geo_ftp.suppl.URL}
\title{GEO raw data from ftp}
\usage{
geo_ftp.suppl.URL(...)
}
\arguments{
\item{...}{one or more GDS, GSE, GPL, GSM number}
}
\value{
URLS
}
\description{
GEO raw data from ftp
}
\examples{
geo_ftp_url_suppl('GSE29450')
geo_ftp_url_suppl('GSE29450', 'GPL3480')
}
