% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_ftp.URL.R
\name{geo_ftp.URL}
\alias{geo_ftp.URL}
\title{GEO ftp data url for download}
\usage{
geo_ftp.URL(...)
}
\arguments{
\item{...}{one or more GDS, GSE, GPL, GSM number}
}
\value{
URLS
}
\description{
GEO ftp data url for download
}
\examples{
geo_ftp_url('GSE29450')
geo_ftp_url('GSE29450', 'GPL3480')
}
