% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_gse.series.matrix.URL.R
\name{geo_gse.series.matrix.URL}
\alias{geo_gse.series.matrix.URL}
\title{URL of gse series matrix}
\usage{
geo_gse.series.matrix.URL(...)
}
\arguments{
\item{...}{one or more gse numbers}
}
\value{
series matrix urls
}
\description{
URL of gse series matrix
}
\examples{
geo_gse.series.matrix.URL('gse10245')
}
