% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_Target.R
\name{geo_Target}
\alias{geo_Target}
\title{GEO Target
Get target information from GEO}
\usage{
geo_Target(GSE, FileName = TRUE)
}
\arguments{
\item{GSE}{GSE number}

\item{FileName}{logical, whether to get Filename, default is TRUE}
}
\value{
a dataframe contains Target information
}
\description{
GEO Target
Get target information from GEO
}
\examples{
\donttest{
geo_Target(GSE = 'GSE21034',FileName=FALSE)
}

}
