% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_webpage.URL.R
\name{geo_webpage.URL}
\alias{geo_webpage.URL}
\title{geo webpage url}
\usage{
geo_webpage.URL(...)
}
\arguments{
\item{...}{one or more GDS, GSE, GPL, GSM number}
}
\value{
URLS
}
\description{
geo webpage url
}
\examples{
geo_webpage_url('GSE29450')
geo_webpage_url('GSE29450', 'GPL3480')
}
