% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bm.R
\name{get_bm}
\alias{get_bm}
\title{Retrieves information from the BioMart database}
\usage{
get_bm(mart, attributes, split = "chromosome_name", ..., times = 10L)
}
\arguments{
\item{mart}{object of class Mart, created with the useMart function.}

\item{attributes}{Attributes you want to retrieve.  A possible list of attributes can be retrieved using the function listAttributes.}

\item{split}{The attribute to split the task, since \code{get_bm} use filters to
implement task spliting, user should ensure the split is in the filter
attributes. You can list all filter attributes for a mart by
\link[biomaRt:listFilters]{listFilters}.}

\item{...}{Other arguments passed to \link[biomaRt:getBM]{getBM}.}

\item{times}{The maximal times to try to download data from biomaRt database.}
}
\value{
A \link[data.table:data.table]{data.table}
}
\description{
A modified \link[biomaRt:getBM]{getBM}, which will split our task into pieces
to overcome the network restrictions.
}
\seealso{
\link[biomaRt:getBM]{getBM}
}
