% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytoband-utils.R
\name{get_cytoband}
\alias{get_cytoband}
\title{Get UCSC cytoband data}
\usage{
get_cytoband(x = "hg38", add_arm = TRUE)
}
\arguments{
\item{x}{One of "hg19" or "hg38". "hg38" is derived from AnnotationHub by
record id "AH53178" and "hg19" is by record id "AH53177". Default: "hg38".}

\item{add_arm}{A scalar logical value indicates whether add a column named
"arm" defining the chromosome-arm for each items.}
}
\value{
A \code{\link[GenomicRanges:GRanges-class]{GenomicRanges}} object containing
cytoband informations.
}
\description{
Get UCSC cytoband data
}
