% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-genome.R
\name{get_genome}
\alias{get_genome}
\title{Get available/installed genomes}
\usage{
get_genome(ref_genome)
}
\arguments{
\item{ref_genome}{BSgenome object or name of the installed BSgenome package.
Default: "hg19". Details see "genome" argument of
\link[BSgenome:available.genomes]{getBSgenome}.}
}
\description{
Get available/installed genomes
}
\keyword{internal}
