% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea.R
\name{gggsea}
\alias{gggsea}
\title{GSEA plot that mimic the plot generated by broad institute's GSEA software}
\usage{
gggsea(
  x,
  geneSetID = NULL,
  subplots = 1:3,
  rel_heights = c(1.8, 0.4, 0.8),
  ES_geom = "line",
  add_bar = TRUE,
  ...
)
}
\arguments{
\item{x}{An \code{gseaResult} object returned by clusterProfiler.}

\item{geneSetID}{gene set ID}

\item{subplots}{which subplots to be displayed}

\item{rel_heights}{relative heights of subplots}

\item{ES_geom}{geom for plotting running enrichment score, one of 'line' or
'dot'.}

\item{add_bar}{A bool, whether add a gsea bar in the second plot.}
}
\value{
A \code{patchwork} object.
}
\description{
GSEA plot that mimic the plot generated by broad institute's GSEA software
}
\references{
\url{https://github.com/YuLab-SMU/enrichplot/blob/devel/R/gseaplot.R}
}
