% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_hdp.R
\name{hdp_data}
\alias{hdp_data}
\title{Extract hdp results}
\usage{
hdp_data(x, ...)
}
\arguments{
\item{x}{A \code{HDP} object returned by \link{run_hdp}.}

\item{...}{\itemize{
\item dpindices: Indices of DP nodes to extract. Default: \code{NULL}.
\item sig_active_cutoff: A numeric of the minimal weight to regard a component
as active. Default: \code{0.1}.
\item remove_zero_lower_ci: A scalar logical indicates exposures with
zero lower confidence interval should be removed.
\item cohort_threshold: A numeric of the minimal proportion (if <1L) or number
of samples (if >= 1L) to regard a component as active. Default: \code{0.05}.
}}
}
\value{
A \code{HDP} object with \code{components} and \code{hdpdata} added.
}
\description{
This involves collecting multiple independent HDP sampling chains into an
hdpSampleMulti object using \link[hdp:hdp_multi_chain]{hdp_multi_chain}.
Components are then extracted using
\link[hdp:hdp_extract_components]{hdp_extract_components}.
}
