% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyclone-utils.R
\name{identify_mut_cn}
\alias{identify_mut_cn}
\title{Identify the copy number in mutation position}
\usage{
identify_mut_cn(
  mut_data,
  cnv_data,
  on_sample = NULL,
  on_chr = "chr",
  mut_pos = "pos",
  start_field = "start",
  end_field = "end"
)
}
\arguments{
\item{mut_data}{A data.frame of mutation data with the top 4 columns (omit
column specified in \code{sample_id}) containing "chromosome", "pos", "ref_counts"
and "var_counts", other columns will be omited. Column names don't matter. If
\code{sample_id} is not NULL, the column specified in \code{sample_id} will be matched
in both mut_data and cnv_data.}

\item{cnv_data}{A data.frame of allele-specific CNV data with the top 5
columns containing "chromosome", "start_pos", "end_pos", "major_cn", and
"minor_cn". "sample_id" is optional, details see mut_data. other columns will
be omited. Column names don't matter.}

\item{on_sample}{A string (can be named), specifying the sample column used
to match mut_data and cnv_data. If NULL, all mut_data and cnv_data will be
regarded from the same sample.}

\item{on_chr}{A string (can be named), specifying the chromosome column used
to match mut_data and cnv_data.}

\item{mut_pos}{A string indicating the column names in \code{mut_data} that
contains the variants positions in the chromosome.}

\item{start_field, end_field}{A string indicating the column names in
\code{cnv_data} that contains the start positions and end position of the genomic
ranges.}
}
\value{
A integrated
}
\description{
Identify the copy number in mutation position
}
