% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lumiR_changeNames.R
\name{lumiR_changeNames}
\alias{lumiR_changeNames}
\title{Change column names for lumiR package}
\usage{
lumiR_changeNames(files)
}
\arguments{
\item{files}{txt files}
}
\value{
wrtie down
}
\description{
Change column names for lumiR package
}
