% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeAssayDF.R
\name{makePerFeatureDF}
\alias{makePerFeatureDF}
\alias{makePerFeatureDF,SummarizedExperiment-method}
\alias{makePerFeatureDF,ExpressionSet-method}
\title{Create a per-feature data.frame}
\usage{
makePerFeatureDF(x, ...)

\S4method{makePerFeatureDF}{SummarizedExperiment}(
  x,
  features = TRUE,
  assay.type = NULL,
  use.rowdata = TRUE,
  swap.rownames = NULL,
  check.names = FALSE
)

\S4method{makePerFeatureDF}{ExpressionSet}(
  x,
  features = TRUE,
  assay.type = NULL,
  use.rowdata = TRUE,
  swap.rownames = NULL,
  check.names = FALSE
)
}
\arguments{
\item{x}{A \link[SummarizedExperiment:RangedSummarizedExperiment-class]{SummarizedExperiment}
or \link[Biobase:class.ExpressionSet]{ExpressionSet} object.}

\item{...}{Other arguments passed to specific methods.}

\item{features}{Logical scalar indicating whether feature assay data of \code{x}
should be included. Alternatively, Character or integer vector specifying the
features for which to extract expression profiles across samples.}

\item{assay.type}{String or integer scalar indicating the assay to use to
obtain expression values. Must refer to a matrix-like object with integer or
numeric values.}

\item{use.rowdata}{Logical scalar indicating whether row metadata of \code{x}
should be included. Alternatively, a character or integer vector specifying
the row metadata fields to use.}

\item{swap.rownames}{String or integer specifying the
\link[SummarizedExperiment:SummarizedExperiment-class]{rowData} column containing the features
names. If \code{NULL}, \code{rownames(x)} will be used.}

\item{check.names}{Logical scalar indicating whether column names of the
output should be made syntactically valid and unique.}
}
\value{
A data.frame.
}
\description{
Create a per-feature data.frame (i.e., where each row represents a feature /
gene) from a
\link[SummarizedExperiment:RangedSummarizedExperiment-class]{SummarizedExperiment}, most
typically for creating custom ggplot2 plots.
}
