% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_ids.R
\name{map_ids}
\alias{map_ids}
\alias{map_ids,ANY-method}
\alias{map_ids,SummarizedExperiment-method}
\alias{map_ids,ExpressionSet-method}
\title{Gets the mapped ids (column) for a set of keys that are of a particular
keytype}
\usage{
map_ids(x, ...)

\S4method{map_ids}{ANY}(x, db, column, keytype, ..., decreasing = TRUE)

\S4method{map_ids}{SummarizedExperiment}(
  x,
  db,
  column,
  keytype,
  ...,
  swap_rownames = NULL,
  assay_name = NULL,
  decreasing = TRUE
)

\S4method{map_ids}{ExpressionSet}(
  x,
  db,
  column,
  keytype,
  ...,
  swap_rownames = NULL,
  assay_name = NULL,
  decreasing = TRUE
)
}
\arguments{
\item{x}{A matrix-like entity, such as a
\link[SummarizedExperiment:RangedSummarizedExperiment-class]{SummarizedExperiment} or an
\link[Biobase:class.ExpressionSet]{ExpressionSet} object.}

\item{...}{Other arguments passed to specific methods. For method, additional
arguments passed to \link[AnnotationDbi:AnnotationDb-class]{mapIds}.}

\item{db}{An \link[AnnotationDbi:AnnotationDb-class]{AnnotationDb} object. But in
practice this will mean an object derived from an AnnotationDb object such as
a \code{OrgDb} or \code{ChipDb} object. Alternatively, a named atomic vector can be
used to match \code{keytype} by names.}

\item{column}{The columns or kinds of things that can be retrieved from the
\code{db} database. Multiple column items can be extracted simutaneously.}

\item{keytype}{The keytype that matches the keys (\code{rownames(x)} or column
specified in \code{swap_rownames}) used.}

\item{decreasing}{A bool, Should the sort order be increasing or decreasing?}

\item{swap_rownames}{String or integer specifying the
\link[SummarizedExperiment:SummarizedExperiment-class]{rowData} or \link[Biobase:featureData]{fData} column
containing the features names. If \code{NULL}, \code{rownames(x)} will be used.}

\item{assay_name}{A string or integer scalar indicating which assay in the
\code{x} to calculate the mean value for each row. If \code{NULL}, the \verb{1st} assay (if
x is a \link[SummarizedExperiment:RangedSummarizedExperiment-class]{SummarizedExperiment}
object) or \link[Biobase:exprs]{exprs} (if x is an
\link[Biobase:class.ExpressionSet]{ExpressionSet} object) will be used.}
}
\description{
This function maps the feature names to a specific keytype.
Unmapped features are removed, and any duplicate mapped features are also
removed, retaining only the first occurrence based on the mean values across
all samples.
}
