% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmf_module.R
\name{nmf_modules}
\alias{nmf_modules}
\title{Identify modules across different NMF ranks.}
\usage{
nmf_modules(nmf_list, threshold = 5L)
}
\arguments{
\item{nmf_list}{A named list of NMF (identified by \link[NMF:NMF-package]{NMF}).}

\item{threshold}{Indicates the minimum number of genes required to filter a
module, failing which, it shall not be considered.}
}
\value{
A list, including a curated selection of NMF objects and the idenfied
modules.
}
\description{
This function aims to ascertain the optimal rank for a given NMF object list
and subsequently identify the modules linked to the aforementioned
determination.
}
\references{
Barkley, D., Moncada, R., Pour, M. et al. Cancer cell states recur across
tumor types and form specific interactions with the tumor microenvironment.
Nat Genet 54, 1192–1201 (2022). https://doi.org/10.1038/s41588-022-01141-9
}
\seealso{
\url{https://github.com/yanailab/PanCancer}
}
