% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oncomine.R
\name{oncomine}
\alias{oncomine}
\title{Clear "Oncomine" database
This function should be used with "Save Page WE" plugin in chrome.}
\usage{
oncomine(...)
}
\arguments{
\item{...}{html data}
}
\value{
a dataframe
}
\description{
Clear "Oncomine" database
This function should be used with "Save Page WE" plugin in chrome.
}
\examples{
library(BioMisc)
x1 <- system.file('extdata','Oncomine1.html',package = 'BioMisc')
x2 <- system.file('extdata','Oncomine2.html',package = 'BioMisc')
oncomine(x1,x2)
}
