% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overdispersion.R
\name{overdispersion}
\alias{overdispersion}
\title{Overdispersion significance analysis for geneset}
\usage{
overdispersion(matrix, features, ..., log = FALSE, size = 10^3L, nbin = 25L)
}
\arguments{
\item{matrix}{Normalized gene expression matrix with row is genes and column
is samples.}

\item{features}{The template for gene sampling, considering genes with
comparable expression levels of these genes.}

\item{...}{
  Arguments passed on to \code{\link[BiocSingular:runPCA]{BiocSingular::runPCA}}
  \describe{
    \item{\code{}}{}
  }}

\item{log}{A scalar logical indicates whether do logarithms.}

\item{size}{A non-negative integer giving the number of sampling.}

\item{nbin}{Divides the dataset into approximately \code{nbin} clusters based on
gene expression levels. Subsequently, the sampling process will be applied
within each cluster for the specified features.}
}
\description{
Principal component analysis was performed on the expression of the specified
geneset (\code{features}), and the variance explained by principal component (PC)1
was calculated. For each module, 1,000 random lists of genes with similar
expression levels were generated, and the variance explained by PC1 in those
genesets was calculated. The significance (p) was calculated as the fraction
of random genesets that resulted in a higher PC1 variance than the geneset
itself.
}
\references{
Barkley, D., Moncada, R., Pour, M. et al. Cancer cell states recur across
tumor types and form specific interactions with the tumor microenvironment.
Nat Genet 54, 1192–1201 (2022). https://doi.org/10.1038/s41588-022-01141-9
}
\seealso{
\url{https://github.com/yanailab/PanCancer}
}
