% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netMHCpan.R
\name{parse_netmhcpan}
\alias{parse_netmhcpan}
\title{Parse NetMHCpan result}
\usage{
parse_netmhcpan(file, ids = NULL, rth = 0.5, rlt = 2)
}
\arguments{
\item{file}{A string of the result from NetMHCpan.}

\item{ids}{A character of identifier of the fasta sequence which was input
into NetMHCpan.}

\item{rth}{Rank Threshold for high binding peptides.}

\item{rlt}{Rank Threshold for low binding peptides.}
}
\value{
A \link[data.table:data.table]{data.table}.
}
\description{
Parse NetMHCpan result
}
\details{
BindLevel: (SB: strong binder, WB: weak binder). The peptide will be
identified as a strong binder if the \% Rank is below the specified threshold
for the strong binders (\code{rth}). The peptide will be identified as a weak
binder if the \% Rank is above the threshold of the strong binders (\code{rth})
but below the specified threshold for the weak binders (\code{rlt}).
}
\seealso{
\url{https://services.healthtech.dtu.dk/services/NetMHCIIpan-4.1/}
}
