% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyclone-utils.R
\name{prepare_pyclone}
\alias{prepare_pyclone}
\title{Prepare data for the input of PyClone or PyClone-vi}
\usage{
prepare_pyclone(
  mut_data,
  cnv_data,
  on_sample = NULL,
  normal_cn = 2L,
  pyclone_vi = FALSE,
  purity = 1L,
  error_rate = NULL
)
}
\arguments{
\item{mut_data}{A data.frame of mutation data with the top 4 columns (omit
column specified in \code{sample_id}) containing "chromosome", "pos", "ref_counts"
and "var_counts", other columns will be omited. Column names don't matter. If
\code{sample_id} is not NULL, the column specified in \code{sample_id} will be matched
in both mut_data and cnv_data.}

\item{cnv_data}{A data.frame of allele-specific CNV data with the top 5
columns containing "chromosome", "start_pos", "end_pos", "major_cn", and
"minor_cn". "sample_id" is optional, details see mut_data. other columns will
be omited. Column names don't matter.}

\item{on_sample}{A string (can be named), specifying the sample column used
to match mut_data and cnv_data. If NULL, all mut_data and cnv_data will be
regarded from the same sample.}

\item{normal_cn}{The copy number of the locus in non-malignant cells. This
should generally be 2 except for sex chromosomes in males.}

\item{pyclone_vi}{a logical value indicates whether prepare data for the
input of PyClone-vi. Details see \url{https://github.com/Roth-Lab/pyclone-vi}}

\item{purity}{Only used for
\href{https://github.com/Roth-Lab/pyclone-vi}{PyClone-vi}, the tumour content
(cellularity) of the sample. Default value is 1.0 if column is not present.}

\item{error_rate}{Only used for
\href{https://github.com/Roth-Lab/pyclone-vi}{PyClone-vi}, sequencing error rate.}
}
\description{
Prepare data for the input of PyClone or PyClone-vi
}
\references{
\itemize{
\item \url{https://github.com/Roth-Lab/pyclone}
\item \url{https://github.com/Roth-Lab/pyclone-vi}
\item \url{https://bitbucket.org/sequenzatools/sequenza/src/v2.1.1/R/next.R}
}
}
