% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qlim.R
\name{qlim}
\alias{qlim}
\title{Run QuSAGE on any limma model fit}
\usage{
qlim(dat, fit, coef, geneSets, n.points = 2^14)
}
\arguments{
\item{dat}{Omic data matrix or matrix-like object with rows corresponding to 
probes and columns to samples. Any object that can be processed by 
\code{\link[limma]{getEAWP}} is acceptable.}

\item{fit}{An object of class \code{\link[limma]{MArrayLM}}, as created by a call to
\code{\link[limma]{eBayes}}.}

\item{coef}{Column name or number specifying which coefficient or contrast
of the linear model is of interest.}

\item{geneSets}{Either a named list of pathways to be compared, or a vector of
probe names representing a single gene set.}

\item{n.points}{The number of points at which to sample the convoluted
\emph{t}-distribution. See Details.}
}
\value{
A data frame with the following columns:
\itemize{
  \item{'Pathway'} Name of the pathway.
  \item{'logFC'} Average log2 fold change of genes in the pathway.
  \item{'p.value'} The gene set \emph{p}-value, as calculated using
    \code{pdf.pVal}.
  \item{'q.value'} The false discovery rate, as estimated using Storey's
    \emph{q}-value method.
}
}
\description{
This function is a wrapper for the QuSAGE algorithm, which tests for pathway 
enrichment, designed for easy integration with limma.
}
\details{
\code{qlim} combines the statistical flexibility and empirical Bayes methods of 
\code{limma} with the specificity and sensitivity of the QuSAGE algorithm for 
detecting pathway enrichment. Simulations have shown that this pipeline 
outperforms each package's independent methods for gene set analysis in 
complex experimental designs, namely \code{\link[limma]{camera}} and 
\code{link[qusage]{qgen}}. See Watson & John, forthcoming.

By default \code{n.points} is set to 2^14, or 16,384 points, which will give very
accurate \emph{p}-values in most cases. Sampling at more points will increase the
accuracy of the resulting \emph{p}-values, but will also linearly increase the
amount of time needed to calculate the result. With larger sample sizes, as few
as 1/4 this number of points can be used without seriously affecting the accuracy
of the resulting \emph{p}-values. However, when there are a small number of samples
(i.e., fewer than 8 samples total), the \emph{t}-distribution must be sampled over
a much wider range, and the number of points needed for sampling should be
increased accordingly. It is recommended that when running \code{qlim} with fewer
than 8 samples, the number of points be increased to at least 2^15 or 2^16. It may
also be useful to test higher values of this parameter, as it can often result in
a much more significant \emph{p}-value with small sample sizes.
}
\examples{
# Fit limma model
library(limma)
eset <- matrix(rnorm(5000 * 10), nrow = 5000, ncol = 10)
dimnames(eset) <- list(seq_len(nrow(eset)), 
                       paste0('S', seq_len(ncol(eset))))
clin <- data.frame(treat = rep(c("ctl", "trt"), each = 5),
                   batch = rep(c("b1", "b2"), times = 5),
                      x1 = rnorm(10),
                      x2 = runif(10))
des <- model.matrix(~ treat + batch + x1 + x2, data = clin)
fit <- eBayes(lmFit(eset, des))

# Create list of differentially expressed pathways
geneSets = list()
for (i in 0:10) {
  genes <- ((30 * i) + 1):(30 * (i + 1))
  eset[genes, clin$treat == "trt"] <- eset[genes, clin$treat == "trt"] + rnorm(1)
  geneSets[[paste("Set", i)]] <- genes
}

# Run qlim
qlim(eset, fit, coef = 2, geneSets)

}
\references{
Smyth, G.K. (2004). "Linear models and empirical Bayes methods for assessing
differential expression in microarray experiments." \emph{Stat. Appl. Genet. 
Molec. Biol.}, \emph{3}(1).
\url{http://www.statsci.org/smyth/pubs/ebayes.pdf}

Yaari, G. Bolen, C.R., Thakar, J. & Kleinstein, S.H. (2013). "Quantitative set
analysis for gene expression: a method to quantify gene set differential expression
including gene-gene correlations." \emph{Nucleic Acids Res.}, \emph{41}(18): e170.
\url{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC3794608/}

Turner, J.A., Bolen, C.R. & Blankenship, D.M. (2015). "Quantitative gene set analysis
generalized for repeated measures, confounder adjustment, and continuous covariates."
\emph{BMC Bioinformatics}, \emph{16}(1): 272.
\url{http://bmcbioinformatics.biomedcentral.com/articles/10.1186/s12859-015-0707-9}

Storey, J. & Tibshirani, R. (2003). "Statistical significance for genomewide
studies." \emph{Proc. Natl. Acad. Sci.}, \emph{100}(16): 9440–9445.
\url{http://www.pnas.org/content/100/16/9440.full}
}
