% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qn.bmiq.R
\name{qn.bmiq}
\alias{qn.bmiq}
\title{Preprocess DNA methylation data}
\usage{
qn.bmiq(targets, idats)
}
\arguments{
\item{targets}{A data frame of sample information or a vector of barcodes.}

\item{idats}{Path to the directory containing the relevant .idat files.}
}
\value{
A matrix of filtered and normalized beta values.
}
\description{
This function runs a QN.BMIQ pipeline on raw .idat files.
}
\details{
\code{qn.bmiq} provides wrappers for functions from several packages that
collectively form a complete preprocessing pipeline on raw DNA methylation data.
It begins by filtering autosomal probes by detection \emph{p}-value, beadcount,
alignment redundancy, and known SNPs using the default settings of
\code{ChAMP::champ.load}. It then runs color bias adjustment, background correction,
and quantile normalization, using functions from the \code{lumi} package. Finally,
data are normalized using beta-mixture quantile dilation as implemented by
\code{wateRmelon::BMIQ}. Missing values are imputed using \emph{k}-nearest
neighbors with the default setting of \emph{k} = 10. Zeros are replaced with a very
small value to facilitate logit transformation.

Normalization procedures for methylation data are the subject of much active
research. Over a dozen methods are implemented in various Bioconductor packages.
The QN.BMIQ pipeline has been found to be optimal in terms of both technical
reproducibility (Marabita et al., 2013) and biological comparison with gold
standard whole genome bisulfite sequencing (Wang et al., 2015).
}
\examples{
data()


}
\references{
Morris, T.J., Butcher, L.M., Teschendorff, A.E., Chakravarthy, A.R., Wojdacz, T.K.
& Beck, S. (2014). "ChAMP: 450k Chip Analysis Methylation Pipeline."
\emph{Bioinformatics}, \emph{30}(3): 428-430.
\url{http://doi.org/10.1093/bioinformatics/btt684}

Bolstad, B.M., Irizarry, R.A. Astrand, M. & Speed, T.P. (2003). "A comparison of
normalization methods for high density oligonucleotide array data based on
variance and bias." \emph{Bioinformatics}, \emph{19}(2): 185-193.
\url{https://www.ncbi.nlm.nih.gov/pubmed/12538238}

Teschendorff, A.E., Marabita, F., Lechner, M. Bartlett, T. Tegner, J. Gomez-Cabrero,
D. & Beck, S. (2012). "A beta-mixture quantile normalization method for correcting
probe design bias in Illumina Infinium 450k DNA methylation data."
\emph{Bioinformatics}, \emph{29}(2): 189-196.
\url{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC3546795/}

Marabita, F. et al. (2013). "An evaluation of analysis pipelines for DNA methylation
profiling using the Illumina HumanMethylation450 BeadChip platform."
\emph{Epigenetics}, \emph{8}(3): 333-346.
\url{https://www.ncbi.nlm.nih.gov/pubmed/23422812}

Wang, T. et al. (2015). "A systematic study of normalization methods for Infinium
450k methylation data using whole-genome bisulfite sequencing data."
\emph{Epigenetics}, \emph{10}(7): 662-669.
\url{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC4623491/}
}
