% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_rrho.R
\name{rrho_heatmap}
\alias{rrho_heatmap}
\title{Rank-Rank Hypergeometric Overlap Map heatmap}
\usage{
rrho_heatmap(rrho_obj, labels, col = NULL, ...)
}
\arguments{
\item{rrho_obj}{a object returned by \code{\link[=run_rrho]{run_rrho()}}}

\item{labels}{Not used currently.}

\item{col}{A vector of colors if the color mapping is discrete or a color
mapping function if the matrix is continuous numbers (should be generated by
\code{circlize::colorRamp2}). If the matrix is continuous, the value can also be a
vector of colors so that colors can be interpolated. Pass to \code{ColorMapping}.
For more details and examples, please refer to \link[ComplexHeatmap:Heatmap]{ComplexHeatmap::Heatmap}}

\item{...}{other parameters passed to \link[ComplexHeatmap:Heatmap]{ComplexHeatmap::Heatmap}}
}
\description{
Rank-Rank Hypergeometric Overlap Map heatmap
}
\examples{
n <- 200
sample1 <- rnorm(n)
sample2 <- rnorm(n)
names(sample1) <- names(sample2) <- paste0("gene", seq_len(n))
rrho_res <- biomisc::run_rrho(sample1, sample2, 1)
biomisc::rrho_heatmap(rrho_res)
}
