% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_rrho.R
\name{rrho_sig_spot}
\alias{rrho_sig_spot}
\title{Rank-Rank Hypergeometric Overlap significant spot}
\usage{
rrho_sig_spot(rrho_obj)
}
\arguments{
\item{rrho_obj}{An object returned by \code{\link[=run_rrho]{run_rrho()}}}
}
\value{
A data.frame, of which each row indicates the highest intensity
point.
}
\description{
Rank-Rank Hypergeometric Overlap significant spot
}
\details{
The highest intensity point on the resulting Rank-Rank
Hypergeometric Overlap map corresponds to the pair of rank thresholds where
the observed statistical overlap between the two gene-expression profiles is
the strongest statistically. In other words, the highest intensity point
represents the optimal overlap between the profiles in that this is the
overlap that is least likely to occur by chance. The coordinates of the
highest intensity point are the rank in each experiment above which are the
most statistically significant set of overlapping genes.
}
