% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_rrho.R
\name{rrho_sig_terms}
\alias{rrho_sig_terms}
\title{Rank-Rank Hypergeometric Overlap Test}
\usage{
rrho_sig_terms(rrho_obj, quadrant = c("up-up", "down-down"))
}
\arguments{
\item{rrho_obj}{a object returned by \code{\link[=run_rrho]{run_rrho()}}}

\item{quadrant}{one or more items in \code{c("up-up", "down-down", "up-down", "down-up")}, controls which quadrant of iterms should be extracted.}
}
\value{
a list
}
\description{
This function just extract the significant items based on RRHO analysis
results.
}
