% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_absolute.R
\name{run_absolute}
\alias{run_absolute}
\title{Automate ABSOLUTE calling for multiple samples in parallel ways}
\usage{
run_absolute(
  seg,
  maf = NULL,
  sigma_p = 0,
  max_sigma_h = 0.015,
  min_ploidy = 0.95,
  max_ploidy = 10,
  primary_disease = NA,
  platform = NULL,
  results_dir = "ABSOLUTE",
  max_as_seg_count = 1500,
  max_neg_genome = 0.005,
  max_non_clonal = 0.05,
  copy_num_type = NULL,
  min_mut_af = 0.1
)
}
\arguments{
\item{seg}{a \code{data.frame} containing columns "Chromosome", "Start",
"End", "Num_Probes", "Segment_Mean". If providing multiple samples, \code{seg}
should contain a column "Sample" to identify different samples}

\item{maf}{MAF, default is \code{NULL}, can provided as \code{data.frame}.}

\item{sigma_p}{Provisional value of excess sample level variance used for
mode search. Default: \code{0}}

\item{max_sigma_h}{Maximum value of excess sample level variance (Eq. 6).
Default: \code{0.015}}

\item{min_ploidy}{Minimum ploidy value to consider. Solutions implying lower
ploidy values will be discarded. Default: \code{0.95}}

\item{max_ploidy}{Maximum ploidy value to consider. Solutions implying
greater ploidy values will be discarded. Default: \code{10}}

\item{primary_disease}{Primary disease of the sample. These shoud be in the
values returned by \code{absolute_disease_map()} and can be one or more. Default:
\code{NA}}

\item{platform}{one of \code{"SNP_6.0"}, \code{"Illumina_WES"},
\code{"SNP_250K_STY"}. Default: \code{"SNP_6.0"}}

\item{results_dir}{directory path used to store result files. Default:
\code{"ABSOLUTE"}}

\item{max_as_seg_count}{Maximum number of allelic segments. Samples with a
higher segment count will be flagged as 'failed'. Default: \code{1500}}

\item{max_neg_genome}{Maximum genome fraction that may be modeled as
non-clonal with copy-ratio below that of clonal homozygous deletion.
Solutions implying greater values will be discarded. Default: \code{0.005}}

\item{max_non_clonal}{Maximum genome fraction that may be modeled as
non-clonal (subclonal SCNA). Solutions implying greater values will be
discarded. Default: \code{0.05}}

\item{copy_num_type}{The type of copy number to be handled. Either \code{total} or
\code{allelic}. Currently \code{allelic} must be used for HAPSEG based inputs and
\code{total} for segmentation file based inputs. Default: \code{"total"}}

\item{min_mut_af}{Minimum mutation allelic fraction. Mutations with lower
allelic fractions will be filtered out before analysis. Default: \code{0.1}}
}
\value{
Side effect. \cr \cr All ABSOLUTE called results (see
\code{\link[ABSOLUTE]{RunAbsolute}}) were kept in directory
\code{file.path(results_dir, "RunAbsolute")}. \cr \cr All summarized
results from multiple ABSOLUTE calling (see
\code{\link[ABSOLUTE]{CreateReviewObject}}) were kept in
\code{file.path(results_dir, "CreateReviewObject")}. \cr \cr All reviewed
results (see \code{\link[ABSOLUTE]{ExtractReviewedResults}}) were kept in
\code{ file.path(results_dir, "reviewed")}
}
\description{
This function is modified based on package
\href{https://github.com/ShixiangWang/DoAbsolute}{DoAbsolute} by adding
parallel evaluation and adjusting personal convention
}
\details{
\href{https://www.nature.com/articles/nbt.2203}{ABSOLUTE} is a
famous software developed by Broad Institute. However, the
\code{\link[ABSOLUTE]{RunAbsolute}} function points to estimate one sample
each time and sets no default values. \code{\link{run_absolute}} helps
users set default parameters based on
\href{https://www.genepattern.org/modules/docs/ABSOLUTE}{ABSOLUTE
documentation} (genepattern don't provide \code{ABSOLUTE} module anymore) and
provides an uniform interface to input data easily and run
\code{\link[ABSOLUTE]{RunAbsolute}} parallelly.

More detail about how to analyze ABSOLUTE results please see
\href{https://www.genepattern.org/analyzing-absolute-data}{analyzing-absolute-data}.
}
\section{Warnings}{

As from R 4.2.0, a length of 2 or more won't be allowed in a \code{if}
condition, You can fix these by installing a modified \code{ABSOLUTE} package by
\code{pak::pkg_install("Yunuuuu/ABSOLUTE")}. The offical version can also be
installed by
\code{install.packages("https://software.broadinstitute.org/cancer/cga/sites/default/files/data/tools/absolute/ABSOLUTE_1.0.6.tar.gz", repos = NULL, type = "source")}
}

\examples{
\donttest{
seg <- readRDS(system.file("extdata", "absolute",
    "run_absolute_example_seg.rds",
    package = "biomisc"
))
maf <- readRDS(system.file("extdata", "absolute",
    "run_absolute_example_maf.rds",
    package = "biomisc"
))
run_absolute(
    seg = seg, maf = maf,
    results_dir = file.path(tempdir(), "ABSOLUTE")
)
}
}
\references{
Carter, S., Cibulskis, K., Helman, E. et al. Absolute quantification of
somatic DNA alterations in human cancer. Nat Biotechnol 30, 413–421 (2012).
\url{https://doi.org/10.1038/nbt.2203}
}
\seealso{
\itemize{
\item \url{https://software.broadinstitute.org/cancer/cga/absolute_run}
\item \url{https://www.genepattern.org/analyzing-absolute-data}
\item \url{https://software.broadinstitute.org/cancer/cga/absolute_download}
}
}
\author{
Yun \email{yunyunpp96@outlook.com}
}
