% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_arm_cn.R
\name{run_arm_cnv}
\alias{run_arm_cnv}
\title{Caculate Chromosome-arm-levels copy number variation}
\usage{
run_arm_cnv(
  seg_cnv,
  cnv_col,
  ref_cytoband = "hg38",
  cnv_mode = c("rel", "abs"),
  ...,
  filter_centromere = TRUE,
  ploidy = NULL,
  threshold = 0.9
)
}
\arguments{
\item{seg_cnv}{A \link[GenomicRanges:GRanges-class]{GenomicRanges} obeject with
CNV values in "cnv_col" column of metadata (absolute segment copy number
determined by ABSOLUTE algorithm, which is the sum of allelic copy numbers,
or relative segment copy number defined with -1 meaning del, 0 meaning
neutral and 1 meaning amp) and samples IDs in "sample_id_col" column.}

\item{cnv_col}{A scalar character gives the column containing CNV values.}

\item{ref_cytoband}{is a \link[GenomicRanges:GRanges-class]{GenomicRanges}
obeject containing the Cytoband reference, It can be a scalar character
\code{"hg19"} or \code{"hg38"}, or you can provided a self-defined
\link[GenomicRanges:GRanges-class]{GenomicRanges} obeject. \code{"hg38"} is derived
from AnnotationHub by record id "AH53178" and \code{"hg19"} is by record id
"AH53177". Default: \code{"hg38"}.}

\item{cnv_mode}{One of "rel" and "abs" correspongding to what Shukla, A and
Cohen-Sharir have presented respectively.}

\item{...}{Not used currently.}

\item{filter_centromere}{Whether to include or exclude segments across
centromere. Default: \code{TRUE}.}

\item{ploidy}{The ploidy to define Chromosome arm-level aneuploidy profiling.
Cohen-Sharir uses background ploidy ("abs" cnv_mode) derived from
\link[ABSOLUTE:RunAbsolute]{run_absolute} algorithm.}

\item{threshold}{The fraction to define Chromosome arm-level aneuploidy
profiling, Shukla, A. used 0.9 as the cut-off ("rel" cnv_mode).}
}
\value{
A data.frame containing Chromosome-arm-levels copy number.
}
\description{
Caculate Chromosome-arm-levels copy number variation
}
\examples{
seg_cnv <- readRDS(system.file("extdata", "run_arm_cnv_example_seg_cnv.rds",
    package = "yjtools"
))
arm_cnv_res <- run_arm_cnv(seg_cnv, "CNV", "barcode")
}
\references{
\itemize{
\item Cohen-Sharir, Y., McFarland, J.M., Abdusamad, M. et al. Aneuploidy
renders cancer cells vulnerable to mitotic checkpoint inhibition. Nature
590, 486–491 (2021).  \url{https://doi.org/10.1038/s41586-020-03114-6}
\item Shukla, A., Nguyen, T.H.M., Moka, S.B. et al. Chromosome arm
aneuploidies shape tumour evolution and drug response. Nat Commun 11, 449
(2020).  \url{https://doi.org/10.1038/s41467-020-14286-0}
}
}
\author{
Yun \email{yunyunpp96@outlook.com}
}
