% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_ccf.R
\name{run_ccf}
\alias{run_ccf}
\title{Matching  mutation Copy number and Estimating CCF (absolute and phylogenetic)}
\usage{
run_ccf(
  mut_data,
  cnv_data,
  on_sample = NULL,
  on_chr = "chr",
  mut_pos = "pos",
  start_field = "startpos",
  end_field = "endpos",
  ...
)
}
\arguments{
\item{mut_data}{A data.frame of mutation data.}

\item{cnv_data}{A data.frame of allele-specific CNV data.}

\item{on_sample}{A string (can be named), specifying the sample column used
to match mut_data and cnv_data. If NULL, all mut_data and cnv_data will be
regarded from the same sample.}

\item{on_chr}{A string (can be named), specifying the chromosome column used
to match mut_data and cnv_data.}

\item{mut_pos}{A string indicating the column names in \code{mut_data} that
contains the variants positions in the chromosome.}

\item{start_field, end_field}{A string indicating the column names in
\code{cnv_data} that contains the start positions and end position of the genomic
ranges.}

\item{...}{
  Arguments passed on to \code{\link[=estimate_ccf]{estimate_ccf}}
  \describe{
    \item{\code{subclone_metric}}{A string, "ccf" or "subclone_prop" specifying how to
choose subclone.}
    \item{\code{min_subclonal}}{Minimal copy number to define subclone.}
    \item{\code{subclone_correction}}{A scalar logical indicates whether subclonal copy
number correction be used.}
    \item{\code{pvalue_correction}}{The method of multiple testing correction be applied
for the copy number correcting mutations. If NULL, no multiple testing
correction will be used.}
    \item{\code{min_vaf_to_explain}}{A numeric, the minimal vaf value to define
subclone.}
  }}
}
\description{
Matching  mutation Copy number and Estimating CCF (absolute and phylogenetic)
}
