% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_hdp.R
\name{run_hdp}
\alias{run_hdp}
\title{Posterior sampling chain across activated DPs.}
\usage{
run_hdp(
  matrix,
  priors = NULL,
  prior_pseudoc = NULL,
  dp_tree_layer = NULL,
  initcc = 10L,
  ...,
  n_posterior = 15L,
  seed = 1234L
)
}
\arguments{
\item{matrix}{A matrix of counts with one row for every sample (same order
as dpindex) and one column for every data category.}

\item{priors}{A matrix of prior distributions (columns must each sum to 1,
number of rows matches number of data categories). If not NULL, the number
of rows must match the number of columns in matrix.}

\item{prior_pseudoc}{Vector of pseudocounts contributed by each prior
distribution. Only used when priors are provided. The frozen nodes
signifying prior components contain some number of pseudo-data items - use
these pseudocounts to specify the relative weighting of each prior
component compared to the volume of observed data. If NULL, will be
\code{rep_len(1000L, ncol(priors))}.}

\item{dp_tree_layer}{A data.frame specifying the children-parent
relationships. The first column must contain all the samples specified in
\code{matrix} (rownames), and the second define the parent, the third difine the
grandparent and more ancestors et.al.}

\item{initcc}{Number of data clusters to start with (every data item is
randomly assigned to a cluster to start with). Will coerced to integer. See
\link[hdp:dp_activate]{dp_activate}.}

\item{...}{
  Arguments passed on to \code{\link[hdp:hdp_posterior]{hdp::hdp_posterior}}
  \describe{
    \item{\code{burnin}}{The number of burn-in iterations.}
    \item{\code{n}}{The number of posterior samples to collect.}
    \item{\code{space}}{The number of iterations between collected samples.}
    \item{\code{cpiter}}{The number of iterations of concentration parameter sampling to perform after each iteration.}
    \item{\code{verbosity}}{Verbosity of debugging statements.
0 (least verbose) -- 4 (most verbose). 0 highly recommended - only change for debugging small examples.}
  }}

\item{n_posterior}{Specifying the number of independent posterior sampling
chains to run. Will coerced to integer.}

\item{seed}{The seed that can be set to reproduce output.}
}
\value{
A list of \link[hdp:hdpSampleChain-class]{hdp::hdpSampleChain} object with the salient
information from each posterior sample.
}
\description{
Run a Gibbs sampler over the activated DP nodes of a Hierarchichal Dirichlet
Process.  Each iteration re-assigns the cluster allocation of every data
item.  Run \code{burnin} iterations, and then collect \code{n} samples from the chain
with \code{space} iterations between each collected sample. Can collect multiple
independent HDP sampling chains in a hdpSampleMulti object via
\link[hdp:hdp_multi_chain]{hdp_multi_chain}. Components are extracted via
\link[hdp:hdp_extract_components]{hdp_extract_components}.
}
\examples{
run_hdp(
    matrix = hdp::example_data_hdp_prior,
    priors = hdp::example_known_priors,
    burnin = 10L, # 10000L
    n = 10L, # 100L
    space = 200L
)
}
\seealso{
\link[hdp:hdp_posterior]{hdp_posterior}
}
