% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_immucellai.R
\name{run_immucellai}
\alias{run_immucellai}
\title{Immune Cell Abundance Identifier}
\usage{
run_immucellai(sample_exp, data_type = c("microarray", "rnaseq"))
}
\arguments{
\item{sample_exp}{Gene expression matrix}

\item{data_type}{One of "rnaseq" and "microarray"}
}
\value{
a data.frame
}
\description{
ImmuCellAI (Immune Cell Abundance Identifier,
\url{http://bioinfo.life.hust.edu.cn/ImmuCellAI/}) is a tool to estimate the
abundance of 24 immune cells from gene expression dataset including RNA-Seq
and microarray data
}
\examples{
sample_exp <- readRDS(system.file(
    "extdata", "run_immucellai_sample_exp.rds",
    package = "biomisc"
))
run_immucellai(sample_exp, "microarray")
}
