% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_motif_fisher.R
\name{run_motif_fisher}
\alias{run_motif_fisher}
\title{Extract single 5' and 3' bases flanking the mutated site for de-novo
signature analysis. Also estimates APOBEC enrichment scores.}
\usage{
run_motif_fisher(
  mut_data,
  ref_genome = NULL,
  signature_motif = c("TCA", "TCT"),
  signature_mut_bases = NULL,
  bg_extension = 20L
)
}
\arguments{
\item{mut_data}{An data.frame with 5 columns in the following order (column
names doesn't matter).
\itemize{
\item sample: sample names
\item chromosome: The variant affected chromosome (chr1).
\item pos: mutation position in the chromosome
\item ref: The plus strand reference allele at this position.
\item alt: Tumor variant allele at this position.
}}

\item{ref_genome}{BSgenome object or name of the installed BSgenome package.
Example: "BSgenome.Hsapiens.UCSC.hg19".}

\item{signature_motif}{A character of the trinucleotide which will be used to
calculate MUTATIONS(tCw) and CONTEXT(tCw) to test for enrichment. Defaul is
APOBEC signature motif: c("TCA", TCT"). All reverseComplemented motif will
also be included.}

\item{signature_mut_bases}{A character of the mutated DNA bases, which will
be used to calculate MUTATIONS(CorG) and CONTEXT(CorG). If NULL, will get
from the second base of signature_motif.}

\item{bg_extension}{The number of bases around the variable base to define
the background. Default 20L.}
}
\value{
A \link[data.table:data.table]{data.table} describing enrichment per
sample.
}
\description{
Extract single 5' and 3' bases flanking the mutated site for de-novo
signature analysis. Also estimates APOBEC enrichment scores.
}
\details{
Extracts immediate 5' and 3' bases flanking the mutated site and
classifies them into 96 substitution classes.
Requires BSgenome data packages for sequence extraction.

APOBEC Enrichment: Enrichment score is calculated using the same method
described by Roberts et al.

E = (MUTATIONS(tCw) * CONTEXT(CorG)) / (MUTATIONS(CorG) * CONTEXT(tCw))

where, MUTATIONS(tCw) = number of mutations within T[C>T]W and T[C>G]W
context. (W -> A or T)

\if{html}{\out{<div class="sourceCode">}}\preformatted{    MUTATIONS(CorG) = number of mutated C and G
}\if{html}{\out{</div>}}

One-sided Fisher's Exact test is performed to determine the enrichment of
signature motif over background.
}
\references{
Roberts SA, Lawrence MS, Klimczak LJ, et al. An APOBEC Cytidine
Deaminase Mutagenesis Pattern is Widespread in Human Cancers. Nature
genetics. 2013;45(9):970-976. doi:10.1038/ng.2702.
}
\seealso{
\url{https://github.com/PoisonAlien/maftools/blob/master/R/TrinucleotideMatrix.R}
}
