% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_rrho.R
\name{run_rrho}
\alias{run_rrho}
\title{Rank-Rank Hypergeometric Overlap Test}
\usage{
run_rrho(list1, list2, stepsize = NULL, alternative = NULL, log_base = exp(1L))
}
\arguments{
\item{list1}{a named numeric vector, For differential gene expression, values
are often \code{-log10(P-value) * sign(effect)}.}

\item{list2}{Same as list1}

\item{stepsize}{Controls the resolution of the test: how many items between
any two overlap tests.}

\item{alternative}{one of c("enrichment", "two.sided")}

\item{log_base}{Normally, \code{hypermat} in the results are
transformed by logarithm, this control the logarithm base. Just like the
\code{base} parameter in \link[base:Log]{base::log} function.}
}
\description{
The function tests for significant overlap between two sorted lists using the method in the reference.
}
