% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature_dissection.R
\name{run_signature_dissection}
\alias{run_signature_dissection}
\title{Expectation Maximisation (EM) to divvy up hdp signatures into known cosimic
signatures}
\usage{
run_signature_dissection(
  signature,
  targets,
  prior = NULL,
  cos_sim_threthold = 0.9,
  emfrac_threshold = 0.1,
  maxiter = 1000L,
  alpha_threshold = 1e-05,
  unique = !is.null(prior)
)
}
\arguments{
\item{signature}{A numeric vector with the same length of \code{ncol(targetes)}}

\item{targets}{A signature matrix with sinature in row and data items in
column (othen COSMIC).}

\item{prior}{A string specifying the prior name in targets of current
signature.}

\item{cos_sim_threthold}{If the cosine similarity is below a certain
threshold (0.95 seems to work well), we then apply an expectation
maximisation (EM) algorithm to deconvolute this signature into its
constituents.}

\item{emfrac_threshold}{The minimal fraction returned by EM to rerun EM for
dessection again.}

\item{maxiter}{An integer specifying the maximal iteration times}

\item{alpha_threshold}{The minimal error threthold, smaller, the results will
be more accurate.}

\item{unique}{A scalar logical indicates whether significant (higher than
\code{cos_sim_threthold}) should be unique, we should set to \code{TRUE} for priors
signature.}
}
\description{
the expectation maximization (EM) algorithm was used to identify pairs of
signatures that might explain the observed signature.
}
\details{
i) For known signatures (e.g. when HDP calls something as “SBS 1”)
\itemize{
\item we calculate the cosine similarity between the HDP version of the
signature, and the PCAWG version of the signature.
\item If the cosine similarity is below a certain threshold (0.95 seems to work
well), we then apply an expectation maximisation (EM) algorithm to
deconvolute this signature into its constituents. E.g. the preconditioned
HDP SBS1 goes to 38.5\% SBS1, 47.5\% SBS5, and 15.0\% SBS18. Reassuringly, 38\%
of the mutations that make up the HDP SBS1 are C to T at CpG, so that fits
well.
}

ii) For novel signatures (i.e. ones that Nicola’s algorithm calls as novel
e.g. "SBS N1")
\itemize{
\item We always apply EM to break an HDP signature down into a composite of
PCAWG signatures.
\item We then reconstitute the signature by adding up the PCAWG signatures
in the proportion that EM gives us (e.g. IDN1 is 27\% ID1, 14\% ID2, 20\% ID5
etc). I have tried only using signatures that contribute >10\% of the
mutations to avoid overfitting.
\item We compute the cosine similarity of the reconstituted signature to the
known signature. If that is less than a certain threshold, then we consider
that the signature is truly novel. Otherwise, we break it down into its
constituents and present the data in that way.
}
}
\references{
Frankell, A.M., Dietzen, M., Al Bakir, M. et al. The evolution of lung cancer
and impact of subclonal selection in TRACERx. Nature 616, 525–533 (2023).
https://doi.org/10.1038/s41586-023-05783-5
}
\seealso{
\url{https://github.com/McGranahanLab/HDP_sigExtraction}
}
