% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_ssgsea.R
\name{run_ssgsea}
\alias{run_ssgsea}
\title{Single-sample GSEA}
\usage{
run_ssgsea(
  data_expr,
  gene_set_list,
  perm = 1000L,
  sample_norm_type = c("rank", "log", "log.rank", "none"),
  weight = 0.75,
  min_sz = 1L,
  max_sz = Inf
)
}
\arguments{
\item{data_expr}{gene expression data. For ssGSEA, normalizd exression values
with gene length adjusted were needed.}

\item{gene_set_list}{gene sets can be provided as \code{list} object}

\item{perm}{the number of permutations to calculate NES, if smaller than
\code{1L}, normalized enrichment scores won't be calculated.}

\item{sample_norm_type}{Normalization method applied to expression data.
Supported methods are rank, log.rank, and log.  (Default: rank)}

\item{weight}{Exponential weight employed in calculation of enrichment
scores. The default value of \code{0.75} was selected after extensive
testing. The module authors strongly recommend against changing from
default. (Default: \code{0.75})}

\item{min_sz}{minimal size of geneSet for testing. (Default: 1L)}

\item{max_sz}{maximal size of geneSet for testing. (Default: Inf)}
}
\value{
a \code{list} of projection results for each gene set and each sample
}
\description{
Project each sample within a data set onto a space of gene set enrichment
scores using the ssGSEA projection methodology. According to GenePattern
(modified) and GSVA
}
\references{
\itemize{\item Subramanian A, Tamayo P, Mootha VK, Mukherjee S,
Ebert BL, Gillette MA, Paulovich A, Pomeroy SL, Golub TR, Lander ES,
Mesirov JP. Gene set enrichment analysis: A knowledge-based approach for
interpreting genome-wide expression profiles. PNAS.
2005;102(43):15545-15550. \url{https://doi.org/10.1073/pnas.0506580102}
\item Barbie, D., Tamayo, P., Boehm, J. et al. Systematic RNA interference
reveals that oncogenic KRAS-driven cancers require TBK1. Nature 462,
108–112 (2009). \url{https://doi.org/10.1038/nature08460}}
}
