% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_wgd.R
\name{run_wgd}
\alias{run_wgd}
\title{Infer Whole-genome doubling}
\usage{
run_wgd(
  seg_data,
  sample_field = NULL,
  ...,
  thresholds = 2:3,
  major_cn_field = "major_cn",
  minor_cn_field = NULL,
  CNt_field = NULL,
  ploidy_field = "ploidy",
  perm_times = 10000L
)
}
\arguments{
\item{seg_data}{A \link{data.frame} obeject with following columns
(arguments ends with "_field").}

\item{sample_field}{A string specifying the sample id column in seg_cnv.}

\item{...}{
  Arguments passed on to \code{\link[=summarize_arm]{summarize_arm}}
  \describe{
    \item{\code{chr_field,start_field,end_field}}{A string specifying the column of the
chromosome name, start positions and end positions of the genomic ranges in
seg_data.}
    \item{\code{ref_cytoband}}{A scalar string \code{"hg19"} or \code{"hg38"}, in this way, see
\link{get_cytoband}, or you can provided a self-defined
\link[GenomicRanges:GRanges-class]{GenomicRanges} obeject containing the
Cytoband reference.}
    \item{\code{contigs}}{A character vector specifying the chromosome to summarize.}
    \item{\code{pruning_mode}}{When some of the seqlevels to drop from \code{ref_cytoband}
are in use (i.e. have ranges on them), the ranges on these sequences need to
be removed before the seqlevels can be dropped. We call this pruning. The
pruning_mode argument controls how to prune x. See
\link[GenomeInfoDb:seqinfo]{seqinfo} pruning.mode}
    \item{\code{arm_field}}{A scalar string indicates the column containing the
chromosome arm. If \code{NULL}, the internal will search the first column starts
with "arm" (ignore letter case). Only values of "p", "q", "acen" and "" (no
arm) are supported.}
    \item{\code{arms}}{A character specifying the arms to return.}
  }}

\item{thresholds}{An integer vector specifying the thresholds to define
Whole-genome doubling.}

\item{major_cn_field}{A string specifying the major_cn (allele-specific)
column in seg_cnv. Default: "major_cn". If iminor_cn_field is NULL,
major_cn_field must exist, in this way, wGD will be calculated based on
Bielski's article. See references.}

\item{minor_cn_field, CNt_field, ploidy_field}{A string specifying the minor_cn
(minor allele), total_cn (tumor absoltue total copy number) and ploidy
column in seg_cnv. If minor_cn_field is not NULL, wGD will be calculated
based on Sally M's article. See references.}

\item{perm_times}{An integer specifying the times of simulation genome.}
}
\description{
Infer Whole-genome doubling
}
\references{
\itemize{
\item Bielski, C.M., Zehir, A., Penson, A.V. et al. Genome doubling shapes the
evolution and prognosis of advanced cancers. Nat Genet 50, 1189–1195
(2018).  \url{https://doi.org/10.1038/s41588-018-0165-1}
\item Tolerance of Whole-Genome Doubling Propagates Chromosomal Instability and
Accelerates Cancer Genome Evolution.
\url{https://doi.org/10.1158/2159-8290.CD-13-0285}
}
}
