% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mut_matrix.R
\name{snv_sub_matrix}
\alias{snv_sub_matrix}
\title{Calculates trinucleotide (or more) context fraction}
\usage{
snv_sub_matrix(
  mut_data,
  sample_field = NULL,
  ref_genome = NULL,
  chr_field = "chr",
  mut_pos = "pos",
  ref_field = "ref",
  alt_field = "alt",
  extension = 1L,
  method = NULL,
  contigs = NULL,
  exome_ranges = NULL
)
}
\arguments{
\item{mut_data}{A data.frame of mutation data.}

\item{sample_field}{A string specifying the sample id column in mut_data.}

\item{ref_genome}{BSgenome object or name of the installed BSgenome package.
Default: "hg19". Details see "genome" argument of
\link[BSgenome:available.genomes]{getBSgenome}.}

\item{chr_field, mut_pos}{A string, specifying the chromosome and variants
positions column in the mut_data.}

\item{ref_field, alt_field}{A string specifying the column of reference allele
or variant allele in mut_data.}

\item{extension}{The number of bases, that's extracted upstream and
downstream of the base substitutions. (Default: 1).}

\item{method}{A string, Default NULL, means no adjustment will be applied.
Details see \link{context_sizefactor}}

\item{contigs}{An atomic vector specifying the chromosome to define the  size
factor to normalize the context matrix.}

\item{exome_ranges}{A \link[GenomicRanges:GRanges-class]{GenomicRanges} Object define
exome ranges.}
}
\value{
the mutation context matrix
}
\description{
Determines trinucleotide (or more) context fraction from a mutation counts
list.
}
