% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparcc.R
\name{sparcc}
\alias{sparcc}
\title{Inferring Correlation Networks from Genomic Survey Data}
\usage{
sparcc(data, ..., times = 200L)
}
\arguments{
\item{data}{Community count data matrix}

\item{...}{
  Arguments passed on to \code{\link[SpiecEasi:sparcc]{SpiecEasi::sparcc}}
  \describe{
    \item{\code{iter}}{Number of iterations in the outer loop}
    \item{\code{inner_iter}}{Number of iterations in the inner loop}
    \item{\code{th}}{absolute value of correlations below this threshold are considered zero by the inner SparCC loop.}
  }}

\item{times}{Number of bootstraps}
}
\value{
A list.
}
\description{
A reimplementation of SparCC algorithm (Friedman et Alm 2012,
PLoS Comp Bio, 2012). SparCC is mainly used for calculating correlations in
compositional data.
}
\references{
\url{https://journals.plos.org/ploscompbiol/article?id=10.1371/journal.pcbi.1002687}
}
\seealso{
\link[SpiecEasi:sparcc]{sparcc}
}
