% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_arm.R
\name{summarize_arm}
\alias{summarize_arm}
\title{Caculate Chromosome-arm-levels copy number variation}
\usage{
summarize_arm(
  seg_data,
  sample_field = NULL,
  other_fields = NULL,
  chr_field = "chr",
  start_field = "startpos",
  end_field = "endpos",
  ref_cytoband = "hg38",
  contigs = NULL,
  arm_field = NULL,
  arms = c("p", "q"),
  ...,
  filter_centromere = TRUE,
  pruning_mode = "error"
)
}
\arguments{
\item{seg_data}{A \link{data.frame} obeject with segmented chromosome
data.}

\item{sample_field}{A string indicates the sample Id column in seg_data.}

\item{other_fields}{A character vector indicates other column in seg_data to
keep in the final result.}

\item{chr_field, start_field, end_field}{A string specifying the column of the
chromosome name, start positions and end positions of the genomic ranges in
seg_data.}

\item{ref_cytoband}{A scalar string \code{"hg19"} or \code{"hg38"}, in this way, see
\link{get_cytoband}, or you can provided a self-defined
\link[GenomicRanges:GRanges-class]{GenomicRanges} obeject containing the
Cytoband reference.}

\item{contigs}{A character vector specifying the chromosome to summarize.}

\item{arm_field}{A scalar string indicates the column containing the
chromosome arm. If \code{NULL}, the internal will search the first column starts
with "arm" (ignore letter case). Only values of "p", "q", "acen" and "" (no
arm) are supported.}

\item{arms}{A character specifying the arms to return.}

\item{...}{Not used currently.}

\item{filter_centromere}{Whether to include or exclude segments across
centromere, namely genomic ranges interseted with "acen" arm of
ref_cytoband.  Default: \code{TRUE}.}

\item{pruning_mode}{When some of the seqlevels to drop from \code{ref_cytoband}
are in use (i.e. have ranges on them), the ranges on these sequences need to
be removed before the seqlevels can be dropped. We call this pruning. The
pruning_mode argument controls how to prune x. See
\link[GenomeInfoDb:seqinfo]{seqinfo} pruning.mode}
}
\value{
A \link[data.table:data.table]{data.table} containing Chromosome-arm
level data.
}
\description{
Caculate Chromosome-arm-levels copy number variation
}
\author{
Yun \email{yunyunpp96@outlook.com}
}
