% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomisc-package.R
\docType{package}
\name{biomisc-package}
\alias{biomisc}
\alias{biomisc-package}
\title{biomisc: Miscellaneous functions for Bioinformatics}
\description{
Miscellaneous functions for Bioinformatics.
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
