% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellmarker.R
\name{cellmarker_search}
\alias{cellmarker_search}
\title{Seach CellMarker database
Details see http://xteam.xbio.top/CellMarker/index.jsp}
\usage{
cellmarker_search(markers, species = "human", internal = NULL)
}
\arguments{
\item{markers}{an atomic character, the markers to search in the CellMarker
database, can be the Gene Symbol, Gene ID, Protein Symbol or Protein ID
(usually starts with "P", "Q" or "O").}

\item{species}{a scalar string, "human" or "mouse".}

\item{internal}{logical value, indicates whether to use internal CellMarker
data. If \code{NULL}, this will be determined automatically; if the CellMarker
data has been downloaded once, namyly we have already used this function once
with a \code{internal} value \code{FALSE}, then the \code{NULL} will indicate \code{FALSE}.
Otherwise \code{TRUE}. The internal data was downloaded from CellMarker on
2022-12-04.}
}
\value{
a data.frame of the searching results, a column named \code{targeted}
containing the matched markers from CellMarker data.
}
\description{
Seach CellMarker database
Details see http://xteam.xbio.top/CellMarker/index.jsp
}
