% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnmf.R
\name{cnmf}
\alias{cnmf}
\title{Modified cNMF}
\usage{
cnmf(
  matrix,
  min_fraction = 0.002,
  k = 15L,
  n_iters = 100L,
  rho = 0.3,
  min_dist = 0.03,
  silhouette = TRUE
)
}
\arguments{
\item{matrix}{A matrix with row is features and column is samples.}

\item{min_fraction}{Only featuers with value (above 0) in at least
\code{min_fraction} of samples are used in cNMF.}

\item{k}{Number of programs.}

\item{n_iters}{The number of times to run NMF internally before making the
consensus}

\item{rho}{Determines number of neighbors to use for calculating KNN distance
as rho * n_iters. can be thought of the fraction of replicates that must
yield a component approximately matching a program in order for that program
to be kept by cNMF.}

\item{min_dist}{distance threshold that determines how close a component must
be to its nearest neighbors in Euclidean space to be considered
‘approximately matching’.}

\item{silhouette}{If TRUE, will calculate silhouette score.}
}
\value{
A list.
}
\description{
A modified version of the cNMF algorithm, implemented in R
}
\references{
\itemize{
\item Simmons, S.K., Lithwick-Yanai, G., Adiconis, X. et al. Mostly natural
sequencing-by-synthesis for scRNA-seq using Ultima sequencing. Nat
Biotechnol 41, 204–211 (2023). https://doi.org/10.1038/s41587-022-01452-6
\item Dylan KotliarAdrian VeresM Aurel NagyShervin TabriziEran HodisDouglas A
MeltonPardis C Sabeti (2019) Identifying gene expression programs of
cell-type identity and cellular activity with single-cell RNA-Seq eLife
8:e43803. https://doi.org/10.7554/eLife.43803
}
}
\seealso{
\itemize{
\item \url{https://github.com/seanken/CompareSequence}
\item \url{https://github.com/dylkot/cNMF}
}
}
