% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_ccf.R
\name{estimate_ccf}
\alias{estimate_ccf}
\title{Estimating CCF (absolute and phylogenetic)}
\usage{
estimate_ccf(
  mut_cn_data,
  sample_field = NULL,
  purity_field = NULL,
  contigs = NULL,
  chr_field = NULL,
  normal_cn = 2L,
  gender,
  min_vaf_to_explain = NULL,
  min_subclonal = NULL,
  conipher = FALSE
)
}
\arguments{
\item{mut_cn_data}{A data.frame with mutation and copy number data. Copy
number often contain subclonal copy number as described in
\href{https://github.com/McGranahanLab/CONIPHER-wrapper/blob/b58235d1cb42d5c7fd54122dc6b9f5e6c4110a75/src/TRACERxHelperFunctions.R#L1}{CONIPHER}.}

\item{sample_field}{A string specifying the sample column. If NULL, all data
will be regarded from the same sample. This is used to confirm every sample
have the same \code{purity}.}

\item{purity_field}{A string specifying the purity column. Default is
"purity".}

\item{contigs}{An atomic vector specifying the chromosome to analyze.}

\item{chr_field}{A string specifying the chromosome column. Only used when
\code{contigs} is not NULL or \code{normal_cn} is NULL. Default is "chr".}

\item{normal_cn}{A scalar number specifying the normal.copy number or a
string indicating the normal_cn column in \code{mut_cn_data}.  It's save to use 2
if you only analyze autosomes. Or you should use \code{gender} to define the
\code{normal_cn}. For sex chromosomes and gender is "male", 1L will be used,
otherwise, 2L will be used.}

\item{gender}{A string, only "female" and "male" are allowed.}

\item{min_vaf_to_explain}{A numeric, the minimal vaf value to define
subclone.}

\item{min_subclonal}{Minimal copy number to define subclone.}

\item{conipher}{A scalar logical indicates whether calculate phyloCCF like
CONIPHER. Details see
\url{https://github.com/McGranahanLab/CONIPHER-wrapper/tree/main}}
}
\description{
Estimating CCF (absolute and phylogenetic)
}
\note{
This function should run for every patient of all region samples.
}
\seealso{
\link{run_ccf}
}
