% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytoband-utils.R
\name{get_arm_ranges}
\alias{get_arm_ranges}
\title{Get arm-level ranges}
\usage{
get_arm_ranges(ref_ranges, arm_col = NULL)
}
\arguments{
\item{ref_ranges}{A \code{\link[GenomicRanges:GRanges-class]{GenomicRanges}} object to
combine into arm-level ranges.}

\item{arm_col}{A scalar string indicates the column containing the chromosome
arm.}
}
\value{
A \code{\link[GenomicRanges:GRanges-class]{GenomicRanges}} object containing
arm-level ranges.
}
\description{
Get arm-level ranges
}
