% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_hdp.R
\name{hdp_data}
\alias{hdp_data}
\title{Extract hdp results}
\usage{
hdp_data(
  hdpsample,
  input_matrix,
  dpindices = NULL,
  incl_comp0 = FALSE,
  sig_active_cutoff = 0.1,
  cohort_threshold = 0.05
)
}
\arguments{
\item{hdpsample}{A hdpSampleChain or hdpSampleMulti}

\item{input_matrix}{A matrix of counts with one row for every sample (same
order as dpindex) and one column for every data category.}

\item{dpindices}{Indices of DP nodes to extract}

\item{incl_comp0}{A scalar logical indicates whether component zero should be
included in the result}

\item{sig_active_cutoff}{A numeric of the minimal weight to regard a
component as active.}

\item{cohort_threshold}{A numeric of the minimal proportion to regard a
component as active.}
}
\description{
Extract hdp results
}
