% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyclone-utils.R
\name{identify_mut_cn}
\alias{identify_mut_cn}
\title{Identify the copy number in mutation position}
\usage{
identify_mut_cn(
  mut_data,
  cnv_data,
  on_sample = NULL,
  on_chr = "chr",
  mut_pos = "pos",
  start_field = "start",
  end_field = "end"
)
}
\arguments{
\item{mut_data}{A data.frame of mutation data.}

\item{cnv_data}{A data.frame of allele-specific CNV data.}

\item{on_sample}{A string (can be named), specifying the sample column used
to match mut_data and cnv_data. If NULL, all mut_data and cnv_data will be
regarded from the same sample.}

\item{on_chr}{A string (can be named), specifying the chromosome column used
to match mut_data and cnv_data.}

\item{mut_pos}{A string indicating the column names in \code{mut_data} that
contains the variants positions in the chromosome.}

\item{start_field, end_field}{A string indicating the column names in
\code{cnv_data} that contains the start positions and end position of the genomic
ranges.}
}
\value{
A integrated data.frame with data column from both mut_data and
cnv_data.
}
\description{
Identify the copy number in mutation position
}
