% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyclone-utils.R
\name{identify_mut_cn}
\alias{identify_mut_cn}
\title{Identify the copy number in mutation position}
\usage{
identify_mut_cn(
  mut_data,
  cnv_data,
  on_sample = NULL,
  on_chr = "chr",
  mut_pos = "pos",
  start_field = "start",
  end_field = "end",
  nomatch = NULL
)
}
\arguments{
\item{mut_data}{A data.frame of mutation data.}

\item{cnv_data}{A data.frame of allele-specific CNV data.}

\item{on_sample}{A string (can be named), specifying the sample column used
to match mut_data and cnv_data. If NULL, all mut_data and cnv_data will be
regarded from the same sample.}

\item{on_chr}{A string (can be named), specifying the chromosome column used
to match mut_data and cnv_data.}

\item{mut_pos}{A string indicating the column names in \code{mut_data} that
contains the variants positions in the chromosome.}

\item{start_field, end_field}{A string indicating the column names in
\code{cnv_data} that contains the start positions and end position of the genomic
ranges.}

\item{nomatch}{When a row in \code{mut_data} has no match to \code{cnv_data},
nomatch=NA means NA is returned. NULL (or 0 for backward compatibility) means
no rows will be returned for that row of \code{mut_data}.}
}
\value{
A integrated \link[data.table:data.table]{data.table} with data column
from both mut_data and cnv_data.
}
\description{
Identify the copy number in mutation position
}
