% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_ids.R
\name{map_ids}
\alias{map_ids}
\alias{map_ids,ANY-method}
\alias{map_ids,SummarizedExperiment-method}
\title{Gets the mapped ids (column) for a set of keys that are of a particular
keytype}
\usage{
map_ids(x, ...)

\S4method{map_ids}{ANY}(x, db, column, keytype, decreasing = TRUE)

\S4method{map_ids}{SummarizedExperiment}(
  x,
  db,
  column,
  keytype,
  swap_rownames = NULL,
  assay_name = NULL,
  decreasing = TRUE
)
}
\arguments{
\item{x}{A matrix-like object.}

\item{...}{Other arguments passed to specific methods.}

\item{db}{An \link[AnnotationDbi:AnnotationDb-class]{AnnotationDb} object. But in
practice this will mean an object derived from an AnnotationDb object such as
a \code{OrgDb} or \code{ChipDb} object.}

\item{column}{The columns or kinds of things that can be retrieved from the
database.}

\item{keytype}{The keytype that matches the keys (\code{rownames(x)}) used.}

\item{decreasing}{A bool, Should the sort order be increasing or decreasing?}

\item{swap_rownames}{String or integer specifying the
\link[SummarizedExperiment:SummarizedExperiment-class]{rowData} column containing the features
names. If \code{NULL}, \code{rownames(x)} will be used.}

\item{assay_name}{A string or integer scalar indicating which
\link[SummarizedExperiment:SummarizedExperiment-class]{assay} in the \code{x} to calculate the mean value
for each row. IF \code{NULL}, the first assay will be used.}
}
\description{
This function maps the feature names to a specific keytype.
Unmapped features are removed, and any duplicate mapped features are also
removed, retaining only the first occurrence based on the mean values across
all samples.
}
