% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmf_consensus.R
\name{nmf_consensus}
\alias{nmf_consensus}
\title{Identify consensus modules.}
\usage{
nmf_consensus(
  basis_list,
  module_size = 50L,
  min_contribution = 0.02,
  min_size = 3L,
  min_intra_sim = 0.7,
  min_intra_size = NULL,
  min_inter_sim = 0.2,
  min_inter_size = 2L,
  redundant_sim = 0.2,
  min_intersection = 0.2,
  founder_intersection_size = 0.2,
  child_intersection_size = 0.2
)
}
\arguments{
\item{basis_list}{A list of sub-list of matrix for each NMF run.}

\item{module_size}{Expected size of genes in a module.}

\item{min_contribution}{Minimal contricution percent (defined by \code{basis}) of
genes in a module.}

\item{min_size}{Minimal number of genes in a module.}

\item{min_intra_sim}{Minimal intra-tumor jaccard index, robust within the
tumour (a program that is represented by several similar NMF programs, as
defined for the same tumour when analysed by multiple NMF rank values; two
NMF programs were considered as similar if they had at least 75\% gene overlap
defined by jaccard index)}

\item{min_intra_size}{Minimal number of significant programs (with jaccard
index > min_intra_sim) to be considered robust within the tumor. If NULL, the
number of NMF runs is used.}

\item{min_inter_sim}{Minimal inter-tumor jaccard index. robust across tumours
(NMF programs that had at least 20\% similarity (by top 50 genes) with any NMF
program in any of the other tumours analysed).}

\item{min_inter_size}{Minimal number of significant programs (with jaccard
index > min_inter_sim) to be considered robust across tuomor.}

\item{redundant_sim}{Non-redundant within the tumour (within each tumour, NMF
programs were ranked by their similarity (gene overlap) with NMFs from other
tumours and selected in decreasing order; once an NMF was selected, any other
NMF within the tumour that had 20\% overlap (or more) with the selected NMF
was removed, to avoid redundancy.}

\item{min_intersection}{Minimal size of gene intersection with other programs
(across all tumor) to be considered significant.}

\item{founder_intersection_size}{Minimal size of significant intersected
programs to define the first NMF program in a cluster.}

\item{child_intersection_size}{Minimal intersection cutoff for adding a new
NMF program to the forming cluster.}
}
\description{
Identify consensus modules.
}
\references{
Gavish, A., Tyler, M., Greenwald, A.C. et al. Hallmarks of transcriptional
intratumour heterogeneity across a thousand tumours. Nature 618, 598–606
(2023). https://doi.org/10.1038/s41586-023-06130-4
}
\seealso{
\url{https://github.com/tiroshlab/3ca}
}
