% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytoband-utils.R
\name{prepare_pyclone}
\alias{prepare_pyclone}
\title{Prepare data for the input of PyClone or PyClone-vi}
\usage{
prepare_pyclone(
  mut_data,
  cnv_data,
  normal_cn = 2,
  pyclone_vi = FALSE,
  sample_id = NULL,
  purity = 1L,
  error_rate = NULL
)
}
\arguments{
\item{mut_data}{a data.frame of mutation data with the top four columns
containing "chromosome", "pos", "ref_counts" and "var_counts", other columns
will be omited. Column names don't matter.}

\item{cnv_data}{a data.frame of allelic CNV data with the top five columns
containing "chromosome", "start_pos", "end_pos", "major_cn", and "minor_cn",
other columns will be omited. Column names don't matter.}

\item{normal_cn}{The copy number of the locus in non-malignant cells. This
should generally be 2 except for sex chromosomes in males.}

\item{pyclone_vi}{a logical value indicates whether prepare data for the
input of PyClone-vi. Details see \url{https://github.com/Roth-Lab/pyclone-vi}}

\item{sample_id}{Only used for
\href{https://github.com/Roth-Lab/pyclone-vi}{PyClone-vi}, the sample names.}

\item{purity}{Only used for
\href{https://github.com/Roth-Lab/pyclone-vi}{PyClone-vi}, the tumour content
(cellularity) of the sample. Default value is 1.0 if column is not present.}

\item{error_rate}{Only used for
\href{https://github.com/Roth-Lab/pyclone-vi}{PyClone-vi}, sequencing error rate.}
}
\description{
Prepare data for the input of PyClone or PyClone-vi
}
\references{
\itemize{
\item \url{https://github.com/Roth-Lab/pyclone}
\item \url{https://github.com/Roth-Lab/pyclone-vi}
}
}
