% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qlim.R
\name{qlim}
\alias{qlim}
\title{Run QuSAGE on any limma model fit}
\usage{
qlim(mat, fit, coef, geneSets, n.points = 2^14)
}
\arguments{
\item{mat}{Omic data matrix with rows of features and columns of samples.
Rownames must match those stored in \code{fit}, as well as the probe names
passed to \code{geneSets}.}

\item{fit}{An object of class \code{MArrayLM}, as created by a call to
\code{limma::lmFit} or \code{limma::eBayes}.}

\item{coef}{Column name specifying which coefficient or contrast
of the linear model is of interest.}

\item{geneSets}{Either a named list of pathways to be compared, or a vector of
probe names representing a single gene set.}

\item{n.points}{The number of points at which to sample the convoluted
\emph{t}-distribution. See Details.}
}
\value{
A data frame with the following columns:
\describe{
  \item{One}{Pathway} Name of the pathway.
  \item{Two}{logFC} Average log2 fold change of genes in the pathway.
  \item{Three}{p.value} The gene set \emph{p}-value, as calculated using
    \code{pdf.pVal}.
  \item{Four}{q.value} The false discovery rate, as estimated using Storey's
    method.
}
}
\description{
This function is a wrapper for the QuSAGE algorithm, which tests for enrichment
across pathways or gene sets.
}
\details{
\code{qlim} extends the methodology of \code{qusage::qgen} to studies of
arbitrary design complexity by operating directly on the output of a \code{limma}
model fit. Taking advantage of that package's extremely flexible statistical
framework, which allows for continuous and/or categorical covariates, probe and/or
sample weights, intra-class correlation estimates, and empirical Bayes shrinkage,
\code{qlim} can take the results of any omic experiment analyzable in \code{limma}
and produce a table of enriched pathways for any given coefficient. See \pkg{limma}
and \pkg{qusage} for more details.

By default \code{n.points} is set to 2^14, or 16,384 points, which will give very
accurate \emph{p}-values in most cases. Sampling at more points will increase the
accuracy of the resulting \emph{p}-values, but will also linearly increase the
amount of time needed to calculate the result. With larger sample sizes, as few
as 1/4 this number of points can be used without seriously affecting the accuracy
of the resulting \emph{p}-values, however, when there are a small number of samples
(i.e., fewer than 8 samples total), the \emph{t}-distribution must be sampled over
a much wider range, and the number of points needed for sampling should be
increased accordingly. It is recommended that when running \code{qlim} with fewer
than 8 samples the number of points be increased to at least 2^15 or 2^16. It may
also be useful to test higher values of this parameter, as it can often result in
a much more significant \emph{p}-value with small sample sizes.
}
\examples{
library(limma)
data(Nevins)
mat <-

}
\references{
Smyth, G.K. (2004). "Linear models and empirical Bayes methods for assessing
differential expression in microarray experiments." \emph{Statistical Applications
in Genetics and Molecular Biology}, \emph{3}(1).
\url{http://www.statsci.org/smyth/pubs/ebayes.pdf}

Yaari, G. Bolen, C.R., Thakar, J. & Kleinstein, S.H. (2013). "Quantitative set
analysis for gene expression: a method to quantify gene set differential expression
including gene-gene correlations." \emph{Nucleic Acids Research}, \emph{41}(18): e170.
\url{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC3794608/}

Turner, J.A., Bolen, C.R. & Blankenship, D.M. (2015). "Quantitative gene set analysis
generalized for repeated measures, confounder adjustment, and continuous covariates."
\emph{BMC Bioinformatics}, \emph{16}(1): 272.
\url{http://bmcbioinformatics.biomedcentral.com/articles/10.1186/s12859-015-0707-9}

Storey, J. & Tibshirani, R. (2003). "Statistical significance for genomewide
studies." \emph{Proc Natl Acad Sci}, \emph{100}(16): 9440–9445.
\url{http://www.pnas.org/content/100/16/9440.full}
}
