% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_rrho.R
\name{rrho_correct_pval}
\alias{rrho_correct_pval}
\title{Multiple hypothesis corrections for RRHO analysis}
\usage{
rrho_correct_pval(
  rrho_obj,
  method = "BY",
  perm = 200L,
  quadrant = c("up-up", "down-down")
)
}
\arguments{
\item{rrho_obj}{An object returned by \code{\link[=run_rrho]{run_rrho()}}}

\item{method}{A string, "permutation" or other correction method (details
see \code{\link[=p.adjust]{p.adjust()}}) indicates which method to use
to correct multiple hypothesis}

\item{perm}{if use "permutation", indicates how many times should be used.}

\item{quadrant}{the "quadrant" to test significance, usually we want to test
whether the overlapping goes in the same direction (over-enrichment), which
means "up-up" quadrant and "down-down" quadrant. You can specify "all" to
test the overall significance of the RRHO map. Due to the sign convention for
over- or under-enrichment in RRHO design, we can test the over-enrichment of
"down-up" quadrant or/and "up-down" quatrant by testing the under-enrichment
of hotspot significance in "down-up" quadrant or/and "up-down" quatrant, the
Pvalue for which test whether the overlapping goes in the different
direction.}
}
\description{
Computes the significance of the agreements between lists as returned by RRHO
using resampling or by Benjamini-Yekutieli correction.
}
\details{
If the Benjamini-Yekutieli corrected hypergeometric map has most significant
areas of absolute log10(P-value) intensity 15 or greater, then permutations
are likely to be significant below a permutation frequency P-value of 0.01.
When the corrected map maxima are less than 15, perform permutations. If the
two maps for comparison are made from gene lists that are of considerably
different lengths, one can either (a) scale the hypergeometric maps using the
list length correction method to account for length differences before
applying a Benjamini-Yekutieli correction or (b) remake the hypergeometric
maps using only items common to all signatures in the set of comparisons
making sure this restriction is not too limiting on the total gene number.

Areas "up-up" quadrant and "down-down" quadrant correspond to overlapping
items that go in the same direction in both experiments. We compare the
maximum in these areas as a summary statistic to screen through permutation
RRHO maps compared to the true RRHO map. The frequency at which permutation
maps have a higher summary statistic than the true map is defined as the
permutation P-value.
}
\examples{
n <- 200
sample1 <- rnorm(n)
sample2 <- rnorm(n)
names(sample1) <- names(sample2) <- paste0("gene", seq_len(n))
rrho_res <- biomisc::run_rrho(sample1, sample2, 1)
progressr::with_progress(
    biomisc::rrho_correct_pval(rrho_res, "permutation", 10L)
)
}
\seealso{
\url{https://academic.oup.com/nar/article/38/17/e169/1033168#82642617}
\url{https://systems.crump.ucla.edu/rankrank/PlaisierSupplemetaryData-SupplementaryMethods_UsersGuide.pdf}
}
