% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_rrho.R
\name{rrho_correct_pval}
\alias{rrho_correct_pval}
\title{Multiple hypothesis corrections for RRHO analysis}
\usage{
rrho_correct_pval(rrho_obj, method = NULL, perm = 200L)
}
\arguments{
\item{rrho_obj}{a object returned by \code{\link[=run_rrho]{run_rrho()}}}

\item{method}{one of \code{c("BY", "permutation")} indicates which method to use
to correct multiple hypothesis}

\item{perm}{if use "permutation", how many permutation times it uses.}
}
\description{
Computes the significance of the agreements between lists as returned by RRHO
using resampling or by Benjamini-Yekutieli correction.
}
\seealso{
\url{https://academic.oup.com/nar/article/38/17/e169/1033168#82642617}
}
