% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_rrho.R
\name{rrho_sig_items}
\alias{rrho_sig_items}
\alias{print.rrho_sig}
\title{Rank-Rank Hypergeometric Overlap significant items}
\usage{
rrho_sig_items(rrho_obj, quadrant = c("up-up", "down-down"))

\method{print}{rrho_sig}(x, ...)
}
\arguments{
\item{rrho_obj}{An object returned by \code{\link[=run_rrho]{run_rrho()}}}

\item{quadrant}{one or more items in \code{c("up-up", "down-down", "up-down", "down-up")}, controls which quadrant of items should be extracted.}

\item{x}{An object returned by \code{\link[=rrho_sig_items]{rrho_sig_items()}}}

\item{...}{Not used currently}
}
\value{
a list of object for each quadrant with class "rrho_sig" which
contains the following fields:
\describe{
\item{sig_item1}{the sinificant items in list1.}
\item{sig_item2}{the sinificant items in list2.}
\item{common_items}{the intersected items between \code{sig_item1} and
\code{sig_item2}.}
\item{hyper_metric}{the corresponding RRHO metric in the significant
coordinate.}
\item{hyper_pvalue}{the corresponding RRHO pvalue in in the significant
coordinate.}
}
}
\description{
This function just extract the significant items based on RRHO analysis
results.
}
\details{
The highest intensity point on the resulting Rank-Rank
Hypergeometric Overlap map corresponds to the pair of rank thresholds where
the observed statistical overlap between the two gene-expression profiles is
the strongest statistically. In other words, the highest intensity point
represents the optimal overlap between the profiles in that this is the
overlap that is least likely to occur by chance. The coordinates of the
highest intensity point are the rank in each experiment above which are the
most statistically significant set of overlapping genes.
}
\examples{
n <- 200
sample1 <- rnorm(n)
sample2 <- rnorm(n)
names(sample1) <- names(sample2) <- paste0("gene", seq_len(n))
rrho_res <- biomisc::run_rrho(sample1, sample2, 1)
biomisc::rrho_sig_items(rrho_res)
}
\references{
\url{https://academic.oup.com/nar/article/38/17/e169/1033168#82642652}
}
