% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_arm_cnv.R
\name{run_arm_cnv}
\alias{run_arm_cnv}
\title{Caculate Chromosome-arm-levels copy number variation}
\usage{
run_arm_cnv(
  seg_cnv,
  cnv_field,
  ref_cytoband = "hg38",
  arm_field = NULL,
  arms = c("p", "q"),
  cnv_mode = c("rel", "abs"),
  ...,
  filter_centromere = TRUE,
  ploidy = NULL,
  threshold = 0.9
)
}
\arguments{
\item{seg_cnv}{A \link[GenomicRanges:GRanges-class]{GenomicRanges} obeject with
CNV values in "cnv_field" column of metadata (absolute segment copy number
determined by ABSOLUTE algorithm, which is the sum of allelic copy numbers,
or relative segment copy number defined with -1 meaning del, 0 meaning
neutral and 1 meaning amp) and samples IDs in "sample_id_col" column.}

\item{cnv_field}{A scalar character gives the column containing CNV values.}

\item{ref_cytoband}{A \link[GenomicRanges:GRanges-class]{GenomicRanges} obeject
containing the Cytoband reference, It can be a scalar character \code{"hg19"} or
\code{"hg38"}, in this way, see \link{get_cytoband}, or you can provided a
self-defined \link[GenomicRanges:GRanges-class]{GenomicRanges} obeject.}

\item{arm_field}{A scalar string indicates the column containing the
chromosome arm. If \code{NULL}, the internal will search the first column starts
with "arm" (ignore letter case). Only values of "p", "q", "acen" and "" (no
arm) are supported.}

\item{arms}{A character specifying the arms to return.}

\item{cnv_mode}{One of "rel" and "abs" correspongding to what Shukla, A and
Cohen-Sharir have presented respectively.}

\item{...}{Not used currently.}

\item{filter_centromere}{Whether to include or exclude segments across
centromere, namely genomic ranges interseted with "acen" arm of
ref_cytoband.  Default: \code{TRUE}.}

\item{ploidy}{The ploidy to define Chromosome arm-level aneuploidy profiling.
Cohen-Sharir uses background ploidy ("abs" cnv_mode) derived from
\link[ABSOLUTE:RunAbsolute]{run_absolute} algorithm.}

\item{threshold}{The fraction to define Chromosome arm-level aneuploidy
profiling, Shukla, A. used 0.9 as the cut-off ("rel" cnv_mode).}
}
\value{
A \link[data.table:data.table]{data.table} containing
Chromosome-arm-levels copy number.
}
\description{
Caculate Chromosome-arm-levels copy number variation
}
\references{
\itemize{
\item Cohen-Sharir, Y., McFarland, J.M., Abdusamad, M. et al. Aneuploidy
renders cancer cells vulnerable to mitotic checkpoint inhibition. Nature
590, 486–491 (2021).  \url{https://doi.org/10.1038/s41586-020-03114-6}
\item Shukla, A., Nguyen, T.H.M., Moka, S.B. et al. Chromosome arm
aneuploidies shape tumour evolution and drug response. Nat Commun 11, 449
(2020).  \url{https://doi.org/10.1038/s41467-020-14286-0}
}
}
\author{
Yun \email{yunyunpp96@outlook.com}
}
