% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_ccf.R
\name{run_ccf}
\alias{run_ccf}
\title{Matching mutation Copy number and Estimating CCF (absolute and phylogenetic)}
\usage{
run_ccf(
  mut_data,
  cnv_data,
  on_patient = NULL,
  on_sample = NULL,
  purity_field = NULL,
  on_chr = "chr",
  mut_pos = "pos",
  start_field = "startpos",
  end_field = "endpos",
  ref_field = NULL,
  alt_field = NULL,
  indel_field = NULL,
  ...,
  nomatch = NULL,
  kept_cols = NULL
)
}
\arguments{
\item{mut_data}{A data.frame of mutation data.}

\item{cnv_data}{A data.frame of allele-specific CNV data.}

\item{on_patient}{A string (can be named), specifying the patient column used
to match mut_data and cnv_data. If NULL, all mut_data and cnv_data will be
regarded from the same patient.}

\item{on_sample}{A string (can be named), specifying the sample column used
to match mut_data and cnv_data. If NULL, all mut_data and cnv_data will be
regarded from the same sample.}

\item{purity_field}{A string specifying the purity column (in \code{mut_data} or
\code{cnv_data}). Default is "purity".}

\item{on_chr}{A string (can be named), specifying the chromosome column used
to match mut_data and cnv_data.}

\item{mut_pos}{A string indicating the column names in \code{mut_data} that
contains the variants positions in the chromosome.}

\item{start_field, end_field}{A string indicating the column names in
\code{cnv_data} that contains the start positions and end position of the genomic
ranges.}

\item{ref_field, alt_field}{A string specifying the column of reference allele
or variant allele in mut_data to estimate whether a variant is a indel or
not. \code{alt_field} will only be used when \code{indel_field} is NULL. \code{ref_field}
must be specified when \code{indel_field} or \code{alt_field} is not NULL.}

\item{indel_field}{A string specifying a logical column of mut_data which
indicates whether or not a variant is a indel.}

\item{...}{
  Arguments passed on to \code{\link[=estimate_ccf]{estimate_ccf}}
  \describe{
    \item{\code{contigs}}{An atomic vector specifying the chromosome to analyze.}
    \item{\code{chr_field}}{A string specifying the chromosome column. Only used when
\code{contigs} is not NULL or \code{normal_cn} is NULL. Default is "chr".}
    \item{\code{normal_cn}}{A scalar number specifying the normal.copy number or a
string indicating the normal_cn column in \code{mut_cn_data}.  It's save to use 2
if you only analyze autosomes. Or you should use \code{gender} to define the
\code{normal_cn}. For sex chromosomes and gender is "male", 1L will be used,
otherwise, 2L will be used.}
    \item{\code{gender}}{A string, only "female" and "male" are allowed.}
    \item{\code{min_vaf_to_explain}}{A numeric, the minimal vaf value to define
subclone.}
    \item{\code{min_subclonal}}{Minimal copy number to define subclone.}
    \item{\code{conipher}}{A scalar logical indicates whether calculate phyloCCF like
CONIPHER. Details see
\url{https://github.com/McGranahanLab/CONIPHER-wrapper/tree/main}}
  }}

\item{nomatch}{When a row in \code{mut_data} has no match to \code{cnv_data},
nomatch=NA means NA is returned. NULL (or 0 for backward compatibility) means
no rows will be returned for that row of \code{mut_data}.  It's not unusual to
place purity data in cnv_data, in this way, if some mutation cannot match the
cnv data, the purity and copy number value for this mutation would be NA, For
CCF estimation, NA is not allowed. Just set nomatch = NULL to omit these
rows.}

\item{kept_cols}{A character vector specifying the columns in \code{mut_data} or
\code{cnv_data} you want to keep in the results. By default only used column and
created columns will be returned.}
}
\description{
Following methods to define subclone and clone.
\itemize{
\item define clone and subclone based on absCCF or mut_multi_btstr
\item define early or late with Mt or phyloCCF.
}

For CONIPHER anlayis, use min_subclonal = 0.05, conipher = TRUE,
min_vaf_to_explain = 0.05. And you should provide indel_field or both
ref_field and alt_field.
}
\note{
\code{on_sample} argument is just for multi-region from a single patient. You
should always use \code{on_patient} to specify the matched sample for a
single-region protocol if you want to correct indel phyloCCF.
}
\seealso{
\itemize{
\item \url{https://bitbucket.org/nmcgranahan/clonalneoantigenanalysispipeline}
\item \url{https://bitbucket.org/nmcgranahan/pancancerclonality}
\item \url{https://github.com/McGranahanLab/CONIPHER-wrapper/}
}
}
