% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_ccf.R
\name{run_ccf}
\alias{run_ccf}
\title{Matching  mutation Copy number and Estimating CCF (absolute and phylogenetic)}
\usage{
run_ccf(
  mut_data,
  cnv_data,
  on_sample = NULL,
  purity_field = NULL,
  on_chr = "chr",
  mut_pos = "pos",
  start_field = "startpos",
  end_field = "endpos",
  ...,
  nomatch = NULL,
  kept_cols = NULL
)
}
\arguments{
\item{mut_data}{A data.frame of mutation data.}

\item{cnv_data}{A data.frame of allele-specific CNV data.}

\item{on_sample}{A string (can be named), specifying the sample column used
to match mut_data and cnv_data. If NULL, all mut_data and cnv_data will be
regarded from the same sample.}

\item{purity_field}{A string specifying the purity column (in \code{mut_data} or
\code{cnv_data}). Default is "purity".}

\item{on_chr}{A string (can be named), specifying the chromosome column used
to match mut_data and cnv_data.}

\item{mut_pos}{A string indicating the column names in \code{mut_data} that
contains the variants positions in the chromosome.}

\item{start_field, end_field}{A string indicating the column names in
\code{cnv_data} that contains the start positions and end position of the genomic
ranges.}

\item{...}{
  Arguments passed on to \code{\link[=estimate_ccf]{estimate_ccf}}
  \describe{
    \item{\code{contigs}}{An atomic vector specifying the chromosome to analyze.}
    \item{\code{normal_cn}}{A scalar number specifying the normal.copy number or a
string indicating the normal_cn column in \code{mut_cn_data}.  It's save to use 2
if you only analyze autosomes. Or you should use \code{gender_field} to define
the \code{normal_cn}. For sex chromosomes and gender is "male", 1L will be used,
otherwise, 2L will be used.}
    \item{\code{gender_field}}{A string specifying the chromosome column. Only used when
normal_cn is NULL. Default is "gender". Only "female" and "male" are
supported in this column.}
    \item{\code{subclone_prop}}{Minimal subclone proportion to define subclone.}
    \item{\code{min_absCCF_higher}}{Minimal right-side 95\% confidence interval of
absolute CCF value to define subclone.}
    \item{\code{min_subclonal}}{Minimal copy number to define subclone.}
    \item{\code{subclone_correction}}{A scalar logical indicates whether subclonal copy
number correction be used.}
    \item{\code{subclone_pvalue_correction}}{The method of multiple testing correction
be applied for the copy number correcting mutations. If NULL, no multiple
testing correction will be used.}
    \item{\code{min_vaf_to_explain}}{A numeric, the minimal vaf value to define
subclone.}
  }}

\item{nomatch}{When a row in \code{mut_data} has no match to \code{cnv_data},
nomatch=NA means NA is returned. NULL (or 0 for backward compatibility) means
no rows will be returned for that row of \code{mut_data}.  It's not unusual to
place purity data in cnv_data, in this way, if some mutation cannot match the
cnv data, the purity and copy number value for this mutation would be NA, For
CCF estimation, NA is not allowed. Just set nomatch = NULL to omit these
rows.}

\item{kept_cols}{A character vector specifying the columns in \code{mut_data} or
\code{cnv_data} you want to keep in the results. By default only used column and
created columns will be returned.}
}
\description{
Matching  mutation Copy number and Estimating CCF (absolute and phylogenetic)
}
