% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_motif_fisher.R
\name{run_motif_fisher}
\alias{run_motif_fisher}
\title{Extract single 5' and 3' bases flanking the mutated site for de-novo
signature analysis. Also estimates APOBEC enrichment scores.}
\usage{
run_motif_fisher(
  mut_data,
  ref_genome,
  signature_motif = c("TCA", "TCT"),
  background_snv = c("C>T", "C>G"),
  remove_non_background_snv = TRUE,
  bg_extension = 20L
)
}
\arguments{
\item{mut_data}{An data.frame with 5 columns in the following order (column
names doesn't matter).
\itemize{
\item sample: sample names
\item chromosome: The variant affected chromosome (chr1).
\item pos: mutation position in the chromosome
\item ref: The plus strand reference allele at this position.
\item alt: Tumor variant allele at this position.
}}

\item{ref_genome}{BSgenome object or name of the installed BSgenome package.
Example: "hg38". Details see "genome" argument of
\link[BSgenome:available.genomes]{getBSgenome}.}

\item{signature_motif}{A character of the trinucleotide which will be used to
calculate MUTATIONS(tCw) and CONTEXT(tCw) to test for enrichment. Defaul is
APOBEC signature motif: c("TCA", TCT"). The second base will define
background MUTATIONS(CorG) and CONTEXT(CorG). All reverseComplemented motif
will also be included.}

\item{background_snv}{A character specifying the mutation type to restrict
the calculation of MUTATIONS(CorG). If NULL, all mutations found in the
second base of \code{signature_motif} will be included. All Complemented
substitution will also be included. Default: c("C>T", "C>G").}

\item{remove_non_background_snv}{A scalar logical indicates whether mutations
not in \code{background_snv} should be removed before analysis.}

\item{bg_extension}{The number of bases around the variable base to define
the background. Default 20L.}
}
\value{
A \link[data.table:data.table]{data.table} describing enrichment per
sample.
}
\description{
Extract single 5' and 3' bases flanking the mutated site for de-novo
signature analysis. Also estimates APOBEC enrichment scores.
}
\details{
Extracts immediate 5' and 3' bases flanking the mutated site and
classifies them into 96 substitution classes.
Requires BSgenome data packages for sequence extraction.

APOBEC Enrichment: Enrichment score is calculated using the same method
described by Roberts et al.

E = (MUTATIONS(tCw) * CONTEXT(CorG)) / (MUTATIONS(CorG) * CONTEXT(tCw))

where, MUTATIONS(tCw) = number of mutations within T[C>T]W and T[C>G]W
context. (W -> A or T)

\if{html}{\out{<div class="sourceCode">}}\preformatted{    MUTATIONS(CorG) = number of mutated C and G
}\if{html}{\out{</div>}}

One-sided Fisher's Exact test is performed to determine the enrichment of
signature motif over background.

APOBEC_enrich mean the enrichment over random of APOBEC pattern mutations.
This is calculated as: {[(TCW to TGW) + (TCW to TTW)]/[(C to G) + (C to
T)]}/[TCW/C]. The enrichment value > 2, which implies that in such samples
at least 50\% of APOBEC signature mutations have been in fact made by APOBEC
enzyme(s). APOBEC_MutLoad_MinEstimate is the minimum estimate of number of
APOBEC induced mutations in a sample. This estimate is calculated using the
formula: [TCW to TGW + TCW to TTW]×[(APOBEC_enrich-1)/APOBEC_enrich] to
determine the number of APOBEC signature mutations in excess of what would be
expected by random mutagenesis. Calculated values are rounded to the nearest
whole number.
}
\references{
\itemize{
\item Roberts SA, Lawrence MS, Klimczak LJ, et al. An APOBEC Cytidine Deaminase
Mutagenesis Pattern is Widespread in Human Cancers. Nature genetics.
2013;45(9):970-976. doi:10.1038/ng.2702.
\item Wang, S., Jia, M., He, Z. et al. APOBEC3B and APOBEC mutational signature
as potential predictive markers for immunotherapy response in non-small
cell lung cancer. Oncogene 37, 3924–3936 (2018).
https://doi.org/10.1038/s41388-018-0245-9
}
}
\seealso{
\url{https://github.com/PoisonAlien/maftools/blob/master/R/TrinucleotideMatrix.R}
}
