% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_rrho.R
\name{run_rrho}
\alias{run_rrho}
\title{Rank-Rank Hypergeometric Overlap Test}
\usage{
run_rrho(list1, list2, stepsize = NULL, log_base = 10L)
}
\arguments{
\item{list1}{a named numeric vector, For differential gene expression, values
are often \code{-log10(P-value) * sign(effect)}.}

\item{list2}{Same as list1}

\item{stepsize}{Controls the resolution of the test: how many items between
any two overlap tests.}

\item{log_base}{Normally, \code{hyper_metric} in the results are
transformed by logarithm, this control the logarithm base. Just like the
\code{base} parameter in \link[base:Log]{base::log} function. Default: \code{10L}.}
}
\description{
The function tests for significant overlap between two sorted lists using the method in the reference.
}
\references{
\url{https://systems.crump.ucla.edu/rankrank/PlaisierSupplemetaryData-SupplementaryMethods_UsersGuide.pdf}
}
